\name{AMCP-package}
\alias{AMCP-package}
\alias{AMCP}
\alias{amcp}
\docType{package}
\title{
\packageTitle{AMCP}
}
\description{
\packageDescription{AMCP}
}
\details{
This package contains all of the data sets used in Maxwell, Delaney, & Kelley (forthcoming). \emph{Designing experiments and analyzing data: A model comparison perspective} (3rd ed.)
}
\author{
\packageAuthor{AMCP}
Maintainer: \packageMaintainer{AMCP}
}
\references{
Maxwell, S. E., Delaney, H. D., & Kelley, K. (forthcoming). \emph{Designing experiments and
analyzing data: A model comparison perspective} (3rd ed.). Routledge.}
\keyword{ package }

