% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{chapter_10_exercise_14}
\alias{chapter_10_exercise_14}
\alias{Chapter_10_Exercise_14}
\alias{C10E14}
\alias{c10e14}
\title{The data used in Chapter 10, Exercise 14}
\format{
An object of class \code{data.frame} with 63 rows and 4 columns.
}
\source{
\url{https://designingexperiments.com/data/}

Maxwell, S. E., Delaney, H. D., & Kelley, K. (2018). \emph{Designing experiments and
analyzing data: {A} model comparison perspective}. (3rd ed.). New York, NY: Routledge.
}
\usage{
data(chapter_10_exercise_14)
}
\description{
Data from Chapter 10 Exercise 14 of \emph{Designing Experiments and Analyzing Data: A Model Comparison Perspective} (3rd edition; Maxwell, Delaney, & Kelley).
}
\details{
\itemize{
\item Composite.
\item Therapist.
\item Modality.
\item Therapist_w_Modality}
}
\section{Synonym}{

C10E14
}

\examples{
# Load the data
data(chapter_10_exercise_14)

# Or, alternatively load the data as
data(C10E14)

# View the structure
str(chapter_10_exercise_14)

# Brief summary of the data.
summary(chapter_10_exercise_14)

}
\references{
Maxwell, S. E., Delaney, H. D., & Kelley, K. (2018). \emph{Designing experiments and analyzing data:
{A} model comparison perspective} (3rd ed.). New York, NY: Routledge.
}
\author{
Ken Kelley \email{kkelley@nd.edu}
}
\keyword{datasets}
