% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{chapter_10_exercise_7}
\alias{chapter_10_exercise_7}
\alias{Chapter_10_Exercise_7}
\alias{C10E7}
\alias{c10e7}
\title{The data used in Chapter 10, Exercise 7}
\format{
An object of class \code{data.frame} with 45 rows and 3 columns.
}
\source{
\url{https://designingexperiments.com/data/}

Maxwell, S. E., Delaney, H. D., & Kelley, K. (2018). \emph{Designing experiments and
analyzing data: {A} model comparison perspective}. (3rd ed.). New York, NY: Routledge.
}
\usage{
data(chapter_10_exercise_7)
}
\description{
Data from Chapter 10 Exercise 7 of \emph{Designing Experiments and Analyzing Data: A Model Comparison Perspective} (3rd edition; Maxwell, Delaney, & Kelley).
}
\details{
\itemize{
\item Ratings.
\item Therapist.
\item Method.}
}
\section{Synonym}{

C10E7
}

\examples{
# Load the data
data(chapter_10_exercise_7)

# Or, alternatively load the data as
data(C10E7)

# View the structure
str(chapter_10_exercise_7)

# Brief summary of the data.
summary(chapter_10_exercise_7)

}
\references{
Maxwell, S. E., Delaney, H. D., & Kelley, K. (2018). \emph{Designing experiments and analyzing data:
{A} model comparison perspective} (3rd ed.). New York, NY: Routledge.
}
\author{
Ken Kelley \email{kkelley@nd.edu}
}
\keyword{datasets}
