% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{chapter_10_table_9}
\alias{chapter_10_table_9}
\alias{C10T9}
\alias{Chapter_10_Table_9}
\alias{c10t9}
\title{The data used in Chapter 10, Table 9}
\format{
An object of class \code{data.frame} with 24 rows and 3 columns.
}
\source{
\url{https://designingexperiments.com/data/}

Maxwell, S. E., Delaney, H. D., & Kelley, K. (2027). \emph{Designing experiments and
analyzing data: {A} model comparison perspective}. (4th ed.). New York, NY: Routledge.
}
\usage{
data(chapter_10_table_9)
}
\description{
The data used in Chapter 10, Table 9
}
\details{
The data in Table 10.9 is based upon the information from the student therapist example of the random-effects section. Assume that the director of the clinic decides to test for a difference across genders in the general severity ratings that graduate students assign to clients. If three male and three female clinical students are randomly selected to participate, and each is randomly assigned four clients with whom to do an intake interview, then we might obtain data like that shown in Table 10.9. Three of the trainees are males while the other three trainees are females. The trainees are nested within their particular gender.
}
\section{Variables}{

\describe{
\item{Gender}{gender of the clinical-student trainee (two levels: male and female)}
\item{Trainee}{trainee, nested within gender (three trainees per gender)}
\item{Severity}{general severity rating assigned by the trainee to a client}
}
}

\section{Synonym}{

C10T9
}

\examples{
# Load the data
data(chapter_10_table_9)

# Or, alternatively load the data as
data(C10T9)

# View the structure
str(chapter_10_table_9)

}
\references{
Maxwell, S. E., Delaney, H. D., & Kelley, K. (2027). \emph{Designing experiments and analyzing data:
{A} model comparison perspective} (4th ed.). New York, NY: Routledge.
}
\author{
Ken Kelley \email{kkelley@nd.edu}
}
\keyword{datasets}
