% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{chapter_11_exercise_17}
\alias{chapter_11_exercise_17}
\alias{Chapter_11_Exercise_17}
\alias{C11E17}
\alias{c11e17}
\title{The data used in Chapter 11, Exercise 17}
\format{
An object of class \code{data.frame} with 14 rows and 4 columns.
}
\source{
\url{https://designingexperiments.com/data/}

Maxwell, S. E., Delaney, H. D., & Kelley, K. (2018). \emph{Designing experiments and
analyzing data: {A} model comparison perspective}. (3rd ed.). New York, NY: Routledge.
}
\usage{
data(chapter_11_exercise_17)
}
\description{
Data from Chapter 11 Exercise 17 of \emph{Designing Experiments and Analyzing Data: A Model Comparison Perspective} (3rd edition; Maxwell, Delaney, & Kelley).
}
\details{
\itemize{
\item Day1.
\item Day2.
\item Day3.
\item Day4.}
}
\section{Synonym}{

C11E17
}

\examples{
# Load the data
data(chapter_11_exercise_17)

# Or, alternatively load the data as
data(C11E17)

# View the structure
str(chapter_11_exercise_17)

# Brief summary of the data.
summary(chapter_11_exercise_17)

}
\references{
Maxwell, S. E., Delaney, H. D., & Kelley, K. (2018). \emph{Designing experiments and analyzing data:
{A} model comparison perspective} (3rd ed.). New York, NY: Routledge.
}
\author{
Ken Kelley \email{kkelley@nd.edu}
}
\keyword{datasets}
