% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{chapter_11_exercise_21}
\alias{chapter_11_exercise_21}
\alias{Chapter_11_Exercise_21}
\alias{C11E21}
\alias{c11e21}
\title{The data used in Chapter 11, Exercise 21}
\format{
An object of class \code{data.frame} with 42 rows and 3 columns.
}
\source{
\url{https://designingexperiments.com/data/}

Maxwell, S. E., Delaney, H. D., & Kelley, K. (2018). \emph{Designing experiments and
analyzing data: {A} model comparison perspective}. (3rd ed.). New York, NY: Routledge.
}
\usage{
data(chapter_11_exercise_21)
}
\description{
Data from Chapter 11 Exercise 21 of \emph{Designing Experiments and Analyzing Data: A Model Comparison Perspective} (3rd edition; Maxwell, Delaney, & Kelley).
}
\details{
\itemize{
\item Mother.
\item Rater.
\item Warmth.}
}
\section{Synonym}{

C11E21
}

\examples{
# Load the data
data(chapter_11_exercise_21)

# Or, alternatively load the data as
data(C11E21)

# View the structure
str(chapter_11_exercise_21)

# Brief summary of the data.
summary(chapter_11_exercise_21)

}
\references{
Maxwell, S. E., Delaney, H. D., & Kelley, K. (2018). \emph{Designing experiments and analyzing data:
{A} model comparison perspective} (3rd ed.). New York, NY: Routledge.
}
\author{
Ken Kelley \email{kkelley@nd.edu}
}
\keyword{datasets}
