% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{chapter_11_exercise_24}
\alias{chapter_11_exercise_24}
\alias{Chapter_11_Exercise_24}
\alias{C11E24}
\alias{c11e24}
\title{The data used in Chapter 11, Exercise 24}
\format{
An object of class \code{data.frame} with 90 rows and 3 columns.
}
\source{
\url{https://designingexperiments.com/data/}

Maxwell, S. E., Delaney, H. D., & Kelley, K. (2018). \emph{Designing experiments and
analyzing data: {A} model comparison perspective}. (3rd ed.). New York, NY: Routledge.
}
\usage{
data(chapter_11_exercise_24)
}
\description{
Data from Chapter 11 Exercise 24 of \emph{Designing Experiments and Analyzing Data: A Model Comparison Perspective} (3rd edition; Maxwell, Delaney, & Kelley).
}
\details{
\itemize{
\item id.
\item judgement.
\item activity.}
}
\section{Synonym}{

C11E24
}

\examples{
# Load the data
data(chapter_11_exercise_24)

# Or, alternatively load the data as
data(C11E24)

# View the structure
str(chapter_11_exercise_24)

# Brief summary of the data.
summary(chapter_11_exercise_24)

}
\references{
Maxwell, S. E., Delaney, H. D., & Kelley, K. (2018). \emph{Designing experiments and analyzing data:
{A} model comparison perspective} (3rd ed.). New York, NY: Routledge.
}
\author{
Ken Kelley \email{kkelley@nd.edu}
}
\keyword{datasets}
