% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{chapter_11_exercise_3}
\alias{chapter_11_exercise_3}
\alias{Chapter_11_Exercise_3}
\alias{C11E3}
\alias{c11e3}
\title{The data used in Chapter 11, Exercise 3}
\format{
An object of class \code{data.frame} with 5 rows and 4 columns.
}
\source{
\url{https://designingexperiments.com/data/}

Maxwell, S. E., Delaney, H. D., & Kelley, K. (2018). \emph{Designing experiments and
analyzing data: {A} model comparison perspective}. (3rd ed.). New York, NY: Routledge.
}
\usage{
data(chapter_11_exercise_3)
}
\description{
Data from Chapter 11 Exercise 3 of \emph{Designing Experiments and Analyzing Data: A Model Comparison Perspective} (3rd edition; Maxwell, Delaney, & Kelley).
}
\details{
\itemize{
\item Location1.
\item Location2.
\item Location3.
\item Location4.}
}
\section{Synonym}{

C11E3
}

\examples{
# Load the data
data(chapter_11_exercise_3)

# Or, alternatively load the data as
data(C11E3)

# View the structure
str(chapter_11_exercise_3)

# Brief summary of the data.
summary(chapter_11_exercise_3)

}
\references{
Maxwell, S. E., Delaney, H. D., & Kelley, K. (2018). \emph{Designing experiments and analyzing data:
{A} model comparison perspective} (3rd ed.). New York, NY: Routledge.
}
\author{
Ken Kelley \email{kkelley@nd.edu}
}
\keyword{datasets}
