% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{chapter_11_table_1}
\alias{chapter_11_table_1}
\alias{C11T1}
\alias{Chapter_11_Table_1}
\alias{c11t1}
\title{The data used in Chapter 11, Table 1}
\format{
An object of class \code{data.frame} with 6 rows and 2 columns.
}
\source{
\url{https://designingexperiments.com/data/}

Maxwell, S. E., Delaney, H. D., & Kelley, K. (2018). \emph{Designing experiments and
analyzing data: {A} model comparison perspective}. (3rd ed.). New York, NY: Routledge.
}
\usage{
data(chapter_11_table_1)
}
\description{
The data used in Chapter 11, Table 1
}
\details{
Table 11.1 displays the data from the observation of six subjects under two treatment conditions, yielding 12 scores in all on the dependent variable.

For the data set, six individuals were observed under two different conditions. The question of interest is: "does the mean of the scores in Condition 1 differ from the mean of the scores in Condition 2?"
}
\section{Variables}{

\describe{
\item{ycond1}{data from the six subjects under treatment condition 1}
\item{ycond2}{data from the six subjects under treatment condition 2}
}
}

\section{Synonym}{

C11T1
}

\examples{
# Load the data
data(chapter_11_table_1)

# Or, alternatively load the data as
data(C11T1)

# View the structure
str(chapter_11_table_1)

}
\references{
Maxwell, S. E., Delaney, H. D., \& Kelley, K. (2018). \emph{Designing experiments and analyzing data:
{A} model comparison perspective} (3rd ed.). New York, NY: Routledge.
}
\author{
Ken Kelley \email{kkelley@nd.edu}
}
\keyword{datasets}
