% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{chapter_12_exercise_18}
\alias{chapter_12_exercise_18}
\alias{Chapter_12_Exercise_18}
\alias{C12E18}
\alias{c12e18}
\title{The data used in Chapter 12, Exercise 18}
\format{
An object of class \code{data.frame} with 10 rows and 3 columns.
}
\source{
\url{https://designingexperiments.com/data/}

Maxwell, S. E., Delaney, H. D., & Kelley, K. (2018). \emph{Designing experiments and
analyzing data: {A} model comparison perspective}. (3rd ed.). New York, NY: Routledge.
}
\usage{
data(chapter_12_exercise_18)
}
\description{
Data from Chapter 12 Exercise 18 of \emph{Designing Experiments and Analyzing Data: A Model Comparison Perspective} (3rd edition; Maxwell, Delaney, & Kelley).
}
\details{
\itemize{
\item Baseline.
\item Feedback.
\item Group.}
}
\section{Synonym}{

C12E18
}

\examples{
# Load the data
data(chapter_12_exercise_18)

# Or, alternatively load the data as
data(C12E18)

# View the structure
str(chapter_12_exercise_18)

# Brief summary of the data.
summary(chapter_12_exercise_18)

}
\references{
Maxwell, S. E., Delaney, H. D., & Kelley, K. (2018). \emph{Designing experiments and analyzing data:
{A} model comparison perspective} (3rd ed.). New York, NY: Routledge.
}
\author{
Ken Kelley \email{kkelley@nd.edu}
}
\keyword{datasets}
