% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{chapter_12_exercise_21}
\alias{chapter_12_exercise_21}
\alias{Chapter_12_Exercise_21}
\alias{C12E21}
\alias{c12e21}
\title{The data used in Chapter 12, Exercise 21}
\format{
An object of class \code{data.frame} with 36 rows and 4 columns.
}
\source{
\url{https://designingexperiments.com/data/}

Maxwell, S. E., Delaney, H. D., & Kelley, K. (2018). \emph{Designing experiments and
analyzing data: {A} model comparison perspective}. (3rd ed.). New York, NY: Routledge.
}
\usage{
data(chapter_12_exercise_21)
}
\description{
Data from Chapter 12 Exercise 21 of \emph{Designing Experiments and Analyzing Data: A Model Comparison Perspective} (3rd edition; Maxwell, Delaney, & Kelley).
}
\details{
\itemize{
\item Recall.
\item Subject.
\item Passage.
\item DifficultyCondition.}
}
\section{Synonym}{

C12E21
}

\examples{
# Load the data
data(chapter_12_exercise_21)

# Or, alternatively load the data as
data(C12E21)

# View the structure
str(chapter_12_exercise_21)

# Brief summary of the data.
summary(chapter_12_exercise_21)

}
\references{
Maxwell, S. E., Delaney, H. D., & Kelley, K. (2018). \emph{Designing experiments and analyzing data:
{A} model comparison perspective} (3rd ed.). New York, NY: Routledge.
}
\author{
Ken Kelley \email{kkelley@nd.edu}
}
\keyword{datasets}
