% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{chapter_12_exercise_22}
\alias{chapter_12_exercise_22}
\alias{Chapter_12_Exercise_22}
\alias{C12E22}
\alias{c12e22}
\title{The data used in Chapter 12, Exercise 22}
\format{
An object of class \code{data.frame} with 40 rows and 8 columns.
}
\source{
\url{https://designingexperiments.com/data/}

Maxwell, S. E., Delaney, H. D., & Kelley, K. (2027). \emph{Designing experiments and
analyzing data: {A} model comparison perspective}. (4th ed.). New York, NY: Routledge.
}
\usage{
data(chapter_12_exercise_22)
}
\description{
Data from Chapter 12 Exercise 22 of \emph{Designing Experiments and Analyzing Data: A Model Comparison Perspective} (4th edition; Maxwell, Delaney, & Kelley). The Hu, Antony, Creery, Vargas, Bodenhausen, and Paller (2015) study of whether implicit social biases can be reduced during sleep. Forty participants were measured on implicit gender bias and implicit racial bias; during a nap, an auditory cue reactivated the counter-bias training for one of the two biases (the cued bias) but not the other (the uncued bias). The exercise analyzes the four bias scores as a 2 (cuing: cued, uncued) by 2 (time: prenap, postnap) fully within-subjects design; higher (more positive) scores reflect more bias.
}
\section{Variables}{

\describe{
\item{Sample}{data collection: 1 = original sample, 2 = replication sample}
\item{CuedBiasType}{the bias that was cued during the nap for this participant: \code{Gender} or \code{Race} (a between-subjects factor the exercise sets aside)}
\item{Cued_Baseline}{cued-bias score at baseline}
\item{Uncued_Baseline}{uncued-bias score at baseline}
\item{Cued_Prenap}{cued-bias score before the nap}
\item{Cued_Postnap}{cued-bias score after the nap}
\item{Uncued_Prenap}{uncued-bias score before the nap}
\item{Uncued_Postnap}{uncued-bias score after the nap}
}
}

\section{Synonym}{

C12E22
}

\examples{
# Load the data
data(chapter_12_exercise_22)

# Or, alternatively load the data as
data(C12E22)

# View the structure
str(chapter_12_exercise_22)

# Brief summary of the data.
summary(chapter_12_exercise_22)

}
\references{
Hu, X., Antony, J. W., Creery, J. D., Vargas, I. M., Bodenhausen, G. V., & Paller, K. A. (2015). Unlearning implicit social biases during sleep. \emph{Science, 348}, 1013--1015.

Maxwell, S. E., Delaney, H. D., & Kelley, K. (2027). \emph{Designing experiments and analyzing data:
{A} model comparison perspective} (4th ed.). New York, NY: Routledge.
}
\author{
Ken Kelley \email{kkelley@nd.edu}
}
\keyword{datasets}
