% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{chapter_12_exercise_25}
\alias{chapter_12_exercise_25}
\alias{Chapter_12_Exercise_25}
\alias{C12E25}
\alias{c12e25}
\title{The data used in Chapter 12, Exercise 25}
\format{
An object of class \code{data.frame} with 39 rows and 3 columns.
}
\source{
\url{https://designingexperiments.com/data/}

Maxwell, S. E., Delaney, H. D., & Kelley, K. (2027). \emph{Designing experiments and
analyzing data: {A} model comparison perspective}. (4th ed.). New York, NY: Routledge.
}
\usage{
data(chapter_12_exercise_25)
}
\description{
Data from Chapter 12 Exercise 25 of \emph{Designing Experiments and Analyzing Data: A Model Comparison Perspective} (4th edition; Maxwell, Delaney, & Kelley). The Kroes, Tendolkar, van Wingen, van Waarde, Strange, and Fernandez (2014) study of electroconvulsive therapy (ECT) and memory reconsolidation. Patients were shown two emotionally aversive slide-show stories; a week later one story was reactivated before three groups were formed (Group A: ECT, tested 24 hours later; Group B: ECT, tested immediately; Group C: no ECT). Each participant then received a multiple-choice memory test for both the reactivated and the non-reactivated story, giving a 3 (group) by 2 (reactivation) design with group between subjects and reactivation within subjects. Scores are percentage correct. The same study also provides the one-way data of \code{chapter_4_exercise_21}; the \code{reactivated} scores here are those Chapter 4 memory scores.
}
\section{Variables}{

\describe{
\item{cond}{group: 1 = Group A (ECT, tested 24 hours later), 2 = Group B (ECT, tested immediately), 3 = Group C (no ECT)}
\item{reactivated}{percentage correct on the story that was reactivated for the participant}
\item{nonreactivated}{percentage correct on the story that was not reactivated}
}
}

\section{Synonym}{

C12E25
}

\examples{
# Load the data
data(chapter_12_exercise_25)

# Or, alternatively load the data as
data(C12E25)

# View the structure
str(chapter_12_exercise_25)

# Brief summary of the data.
summary(chapter_12_exercise_25)

}
\references{
Kroes, M. C. W., Tendolkar, I., van Wingen, G. A., van Waarde, J. A., Strange, B. A., & Fernandez, G. (2014). An electroconvulsive therapy procedure impairs reconsolidation of episodic memories in humans. \emph{Nature Neuroscience, 17}, 204--206.

Maxwell, S. E., Delaney, H. D., & Kelley, K. (2027). \emph{Designing experiments and analyzing data:
{A} model comparison perspective} (4th ed.). New York, NY: Routledge.
}
\author{
Ken Kelley \email{kkelley@nd.edu}
}
\keyword{datasets}
