% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{chapter_12_exercise_26}
\alias{chapter_12_exercise_26}
\alias{Chapter_12_Exercise_26}
\alias{C12E26}
\alias{c12e26}
\title{The data used in Chapter 12, Exercise 26}
\format{
An object of class \code{data.frame} with 58 rows and 6 columns.
}
\source{
\url{https://designingexperiments.com/data/}

Maxwell, S. E., Delaney, H. D., & Kelley, K. (2026). \emph{Designing experiments and
analyzing data: {A} model comparison perspective}. (4th ed.). New York, NY: Routledge.
}
\usage{
data(chapter_12_exercise_26)
}
\description{
Data from Chapter 12 Exercise 26 of \emph{Designing Experiments and Analyzing Data: A Model Comparison Perspective} (4th edition; Maxwell, Delaney, & Kelley). The Gibson, Radvansky, Johnson, and McNerney (2012) study of whether grapheme-color synesthetes have better memory for word lists than nonsynesthetes. Lists of high-frequency or low-frequency words were presented to 48 nonsynesthetes and 10 synesthetes, and two dependent variables were measured for each word frequency: input-output (IO) correspondence and recall accuracy.
}
\section{Variables}{

\describe{
\item{ID}{participant identifier}
\item{Group}{0 = nonsynesthete (control), 1 = grapheme-color synesthete}
\item{HFcorr}{input-output correspondence for high-frequency words, a proportion measuring the extent to which temporal order was preserved in recall (higher scores reflect greater use of relational cues)}
\item{LFcorr}{input-output correspondence for low-frequency words}
\item{HFrecall}{recall accuracy for high-frequency words (higher scores reflect higher accuracy)}
\item{LFrecall}{recall accuracy for low-frequency words}
}
}

\section{Synonym}{

C12E26
}

\examples{
# Load the data
data(chapter_12_exercise_26)

# Or, alternatively load the data as
data(C12E26)

# View the structure
str(chapter_12_exercise_26)

# Brief summary of the data.
summary(chapter_12_exercise_26)

}
\references{
Gibson, B. S., Radvansky, G. A., Johnson, A. C., & McNerney, M. W. (2012). Grapheme-color synesthesia can enhance immediate memory without disrupting the encoding of relational cues. \emph{Psychonomic Bulletin & Review, 19}, 1172--1177.

Maxwell, S. E., Delaney, H. D., & Kelley, K. (2026). \emph{Designing experiments and analyzing data:
{A} model comparison perspective} (4th ed.). New York, NY: Routledge.
}
\author{
Ken Kelley \email{kkelley@nd.edu}
}
\keyword{datasets}
