% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{chapter_12_exercise_27}
\alias{chapter_12_exercise_27}
\alias{Chapter_12_Exercise_27}
\alias{C12E27}
\alias{c12e27}
\title{The data used in Chapter 12, Exercise 27}
\format{
An object of class \code{data.frame} with 193 rows and 6 columns.
}
\source{
\url{https://designingexperiments.com/data/}

Maxwell, S. E., Delaney, H. D., & Kelley, K. (2026). \emph{Designing experiments and
analyzing data: {A} model comparison perspective}. (4th ed.). New York, NY: Routledge.
}
\usage{
data(chapter_12_exercise_27)
}
\description{
Data from Chapter 12 Exercise 27 of \emph{Designing Experiments and Analyzing Data: A Model Comparison Perspective} (4th edition; Maxwell, Delaney, & Kelley). Reports of the number of negative life events collected from three members (mother, father, and child) of 95 first-marriage nuclear families and 98 stepfamilies, as described by Bray and Maxwell (1995). Family member is a within-subjects factor and type of family is a between-subjects factor.
}
\section{Variables}{

\describe{
\item{id}{family identifier}
\item{grp}{type of family: 1 = nuclear family, 2 = stepfamily}
\item{sex}{sex of the child: 1 = boy, 2 = girl}
\item{mtb}{number of negative life events reported by the mother}
\item{ftb}{number of negative life events reported by the father}
\item{ctb}{number of negative life events reported by the child}
}
}

\section{Synonym}{

C12E27
}

\examples{
# Load the data
data(chapter_12_exercise_27)

# Or, alternatively load the data as
data(C12E27)

# View the structure
str(chapter_12_exercise_27)

# Brief summary of the data.
summary(chapter_12_exercise_27)

}
\references{
Bray, J. H., & Maxwell, S. E. (1995). Multivariate statistics for family psychology research. \emph{Journal of Family Psychology, 9}, 144--160.

Maxwell, S. E., Delaney, H. D., & Kelley, K. (2026). \emph{Designing experiments and analyzing data:
{A} model comparison perspective} (4th ed.). New York, NY: Routledge.
}
\author{
Ken Kelley \email{kkelley@nd.edu}
}
\keyword{datasets}
