% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{chapter_12_table_11}
\alias{chapter_12_table_11}
\alias{C12T11}
\alias{Chapter_12_Table_11}
\alias{c12t11}
\title{The data used in Chapter 12, Table 11}
\format{
An object of class \code{data.frame} with 10 rows and 3 columns.
}
\source{
\url{https://designingexperiments.com/data/}

Maxwell, S. E., Delaney, H. D., & Kelley, K. (2018). \emph{Designing experiments and
analyzing data: {A} model comparison perspective}. (3rd ed.). New York, NY: Routledge.
}
\usage{
data(chapter_12_table_11)
}
\description{
The data used in Chapter 12, Table 11
}
\details{
A third covariance matrix is relevant for the AB interaction because this effect does not average over either A or B. Instead, the interaction assesses whether the B difference is the same at each level of A. Table 12.11 presents scores that address this question. For each subject, a given score represents the noise effect (i.e., reaction time when noise is present minus reaction time when noise is absent) at a particular level of the angle factor.

#' @section Variables:
\describe{
\item{angle0}{reaction time when noise is present minus reaction time when noise is absent at angle 0}
\item{angle4}{reaction time when noise is present minus reaction time when noise is absent at angle 4}
\item{angle8}{reaction time when noise is present minus reaction time when noise is absent at angle 8}
}
}
\section{Synonym}{

C12T11
}

\examples{
# Load the data
data(chapter_12_table_11)

# Or, alternatively load the data as
data(C12T11)

# View the structure
str(chapter_12_table_11)

}
\references{
Maxwell, S. E., Delaney, H. D., \& Kelley, K. (2018). \emph{Designing experiments and analyzing data:
{A} model comparison perspective} (3rd ed.). New York, NY: Routledge.
}
\author{
Ken Kelley \email{kkelley@nd.edu}
}
\keyword{datasets}
