% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{chapter_12_table_15}
\alias{chapter_12_table_15}
\alias{C12T15}
\alias{Chapter_12_Table_15}
\alias{c12t15}
\title{The data used in Chapter 12, Table 15}
\format{
An object of class \code{data.frame} with 10 rows and 3 columns.
}
\source{
\url{https://designingexperiments.com/data/}

Maxwell, S. E., Delaney, H. D., & Kelley, K. (2018). \emph{Designing experiments and
analyzing data: {A} model comparison perspective}. (3rd ed.). New York, NY: Routledge.
}
\usage{
data(chapter_12_table_15)
}
\description{
The data used in Chapter 12, Table 15
}
\details{
The data in Table 12.15 consist of reaction time scores for 10 young participants where each participant contributes 3 scores to the analysis. In particular, each participant is exposed to each of 3 experimental conditions, angle (0, 4, and 8). For the current analyses Table 12.15 is appended to Table 12.7, which contains reaction time scores for 10 old participants for angles of 0, 4, and 8. Thus, it is necessary to perform some data management before analyzing the data.
}
\section{Variables}{

\describe{
\item{angle0}{reaction time when noise is present minus reaction time when noise is absent at angle 0}
\item{angle4}{reaction time when noise is present minus reaction time when noise is absent at angle 4}
\item{angle8}{reaction time when noise is present minus reaction time when noise is absent at angle 8}
}
}

\section{Synonym}{

C12T15
}

\examples{
# Load the data
data(chapter_12_table_15)

# Or, alternatively load the data as
data(C12T15)

# View the structure
str(chapter_12_table_15)

}
\references{
Maxwell, S. E., Delaney, H. D., \& Kelley, K. (2018). \emph{Designing experiments and analyzing data:
{A} model comparison perspective} (3rd ed.). New York, NY: Routledge.
}
\author{
Ken Kelley \email{kkelley@nd.edu}
}
\keyword{datasets}
