% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{chapter_12_table_7}
\alias{chapter_12_table_7}
\alias{C12T7}
\alias{Chapter_12_Table_7}
\alias{c12t7}
\title{The data used in Chapter 12, Table 7}
\format{
An object of class \code{data.frame} with 10 rows and 3 columns.
}
\source{
\url{https://designingexperiments.com/data/}

Maxwell, S. E., Delaney, H. D., & Kelley, K. (2018). \emph{Designing experiments and
analyzing data: {A} model comparison perspective}. (3rd ed.). New York, NY: Routledge.
}
\usage{
data(chapter_12_table_7)
}
\description{
The data used in Chapter 12, Table 7
}
\details{
Table 12.7 presents scores for the individual subjects for the A effect for the data in Table 12.1. Notice that each score for a given subject is simply that subject's mean response time for that angle, where the mean is the average of the noise-absent and the noise-present scores.
}
\section{Variables}{

\describe{
\item{angle0}{mean reaction time for subject at angle factor 0, averaging over noise}
\item{angle4}{mean reaction time for subject at angle factor 4, averaging over noise}
\item{angle8}{mean reaction time for subject at angle factor 8, averaging over noise}
}
}

\section{Synonym}{

C12T7
}

\examples{
# Load the data
data(chapter_12_table_7)

# Or, alternatively load the data as
data(C12T7)

# View the structure
str(chapter_12_table_7)

}
\references{
Maxwell, S. E., Delaney, H. D., \& Kelley, K. (2018). \emph{Designing experiments and analyzing data:
{A} model comparison perspective} (3rd ed.). New York, NY: Routledge.
}
\author{
Ken Kelley \email{kkelley@nd.edu}
}
\keyword{datasets}
