% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{chapter_12_table_9}
\alias{chapter_12_table_9}
\alias{C12T9}
\alias{Chapter_12_Table_9}
\alias{c12t9}
\title{The data used in Chapter 12, Table 9}
\format{
An object of class \code{data.frame} with 10 rows and 2 columns.
}
\source{
\url{https://designingexperiments.com/data/}

Maxwell, S. E., Delaney, H. D., & Kelley, K. (2018). \emph{Designing experiments and
analyzing data: {A} model comparison perspective}. (3rd ed.). New York, NY: Routledge.
}
\usage{
data(chapter_12_table_9)
}
\description{
The data used in Chapter 12, Table 9
}
\details{
A different covariance matrix is relevant for the B main effect because the B effect averages over levels of A, whereas the A effect averages over levels of B. Table 12.9 presents each subject's mean score for noise absent and noise present, where the mean is the average of the three angle scores at that particular level of noise.
}
\section{Variables}{

\describe{
   \item{noiseabs}{mean reaction time for subject without noise, averaging over angle}
   \item{noiseprs}{mean reaction time for subject with noise, averaging over angle}
}
}

\section{Synonym}{

C12T9
}

\examples{
# Load the data
data(chapter_12_table_9)

# Or, alternatively load the data as
data(C12T9)

# View the structure
str(chapter_12_table_9)

}
\references{
Maxwell, S. E., Delaney, H. D., \& Kelley, K. (2018). \emph{Designing experiments and analyzing data:
{A} model comparison perspective} (3rd ed.). New York, NY: Routledge.
}
\author{
Ken Kelley \email{kkelley@nd.edu}
}
\keyword{datasets}
