% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{chapter_13_exercise_10}
\alias{chapter_13_exercise_10}
\alias{Chapter_13_Exercise_10}
\alias{C13E10}
\alias{c13e10}
\title{The data used in Chapter 13, Exercise 10}
\format{
An object of class \code{data.frame} with 14 rows and 4 columns.
}
\source{
\url{https://designingexperiments.com/data/}

Maxwell, S. E., Delaney, H. D., & Kelley, K. (2018). \emph{Designing experiments and
analyzing data: {A} model comparison perspective}. (3rd ed.). New York, NY: Routledge.
}
\usage{
data(chapter_13_exercise_10)
}
\description{
Data from Chapter 13 Exercise 10 of \emph{Designing Experiments and Analyzing Data: A Model Comparison Perspective} (3rd edition; Maxwell, Delaney, & Kelley).
}
\details{
\itemize{
\item Face.
\item Circle.
\item Paper.
\item White.}
}
\section{Synonym}{

C13E10
}

\examples{
# Load the data
data(chapter_13_exercise_10)

# Or, alternatively load the data as
data(C13E10)

# View the structure
str(chapter_13_exercise_10)

# Brief summary of the data.
summary(chapter_13_exercise_10)

}
\references{
Maxwell, S. E., Delaney, H. D., & Kelley, K. (2018). \emph{Designing experiments and analyzing data:
{A} model comparison perspective} (3rd ed.). New York, NY: Routledge.
}
\author{
Ken Kelley \email{kkelley@nd.edu}
}
\keyword{datasets}
