% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{chapter_13_exercise_22}
\alias{chapter_13_exercise_22}
\alias{Chapter_13_Exercise_22}
\alias{C13E22}
\alias{c13e22}
\title{The data used in Chapter 13, Exercise 22}
\format{
An object of class \code{data.frame} with 5 rows and 3 columns.
}
\source{
\url{https://designingexperiments.com/data/}

Maxwell, S. E., Delaney, H. D., & Kelley, K. (2018). \emph{Designing experiments and
analyzing data: {A} model comparison perspective}. (3rd ed.). New York, NY: Routledge.
}
\usage{
data(chapter_13_exercise_22)
}
\description{
Data from Chapter 13 Exercise 22 of \emph{Designing Experiments and Analyzing Data: A Model Comparison Perspective} (3rd edition; Maxwell, Delaney, & Kelley).
}
\details{
\itemize{
\item Condition1.
\item Condition2.
\item Condition3.}
}
\section{Synonym}{

C13E22
}

\examples{
# Load the data
data(chapter_13_exercise_22)

# Or, alternatively load the data as
data(C13E22)

# View the structure
str(chapter_13_exercise_22)

# Brief summary of the data.
summary(chapter_13_exercise_22)

}
\references{
Maxwell, S. E., Delaney, H. D., & Kelley, K. (2018). \emph{Designing experiments and analyzing data:
{A} model comparison perspective} (3rd ed.). New York, NY: Routledge.
}
\author{
Ken Kelley \email{kkelley@nd.edu}
}
\keyword{datasets}
