% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{chapter_13_exercise_25}
\alias{chapter_13_exercise_25}
\alias{Chapter_13_Exercise_25}
\alias{C13E25}
\alias{c13e25}
\title{The data used in Chapter 13, Exercise 25}
\format{
An object of class \code{data.frame} with 30 rows and 3 columns.
}
\source{
\url{https://designingexperiments.com/data/}

Maxwell, S. E., Delaney, H. D., & Kelley, K. (2018). \emph{Designing experiments and
analyzing data: {A} model comparison perspective}. (3rd ed.). New York, NY: Routledge.
}
\usage{
data(chapter_13_exercise_25)
}
\description{
Data from Chapter 13 Exercise 25 of \emph{Designing Experiments and Analyzing Data: A Model Comparison Perspective} (3rd edition; Maxwell, Delaney, & Kelley).
}
\details{
\itemize{
\item self.
\item friend.
\item case.}
}
\section{Synonym}{

C13E25
}

\examples{
# Load the data
data(chapter_13_exercise_25)

# Or, alternatively load the data as
data(C13E25)

# View the structure
str(chapter_13_exercise_25)

# Brief summary of the data.
summary(chapter_13_exercise_25)

}
\references{
Maxwell, S. E., Delaney, H. D., & Kelley, K. (2018). \emph{Designing experiments and analyzing data:
{A} model comparison perspective} (3rd ed.). New York, NY: Routledge.
}
\author{
Ken Kelley \email{kkelley@nd.edu}
}
\keyword{datasets}
