% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{chapter_13_table_10}
\alias{chapter_13_table_10}
\alias{C13T10}
\alias{Chapter_13_Table_10}
\alias{c13t10}
\title{The data used in Chapter 13, Table 10}
\format{
An object of class \code{data.frame} with 8 rows and 2 columns.
}
\source{
\url{https://designingexperiments.com/data/}

Maxwell, S. E., Delaney, H. D., & Kelley, K. (2027). \emph{Designing experiments and
analyzing data: {A} model comparison perspective}. (4th ed.). New York, NY: Routledge.
}
\usage{
data(chapter_13_table_10)
}
\description{
The data used in Chapter 13, Table 10
}
\details{
For the hypothetical data contained in Table 13.2, the linear and quadratic D variables were formed by making use of the appropriate coefficients from Appendix Table A.10. Because the eight participants were measured at three occasions, both a linear and a quadratic effect can be tested. The question of interest in this instance is: "is there a linear and/or quadratic trend exhibited by the group over time?" Recall that in the book it was shown that the D variables for linear and quadratic effects led to an omnibus F test of 19.148, which was a value previously obtained for the omnibus effect. Because the particular values chosen for the D variables do not matter (unless it leads to a linear combination of columns), we will focus only on the tests of the individual contrasts when analyzing the data given in Table 13.10. Because columns one and two already represent the linear and quadratic effect respectively, all that needs to be done is to test the mean of the column in order to determine if it differs from zero.
}
\note{
Renumbered for the 4th edition: in the 3rd edition (AMCP 1.x) these data were Table 13.12 (\code{chapter_13_table_12} / \code{C13T12}). The data are unchanged.
}
\section{Variables}{

\describe{
\item{Linear}{linear-trend D variable: the linear contrast applied to each participant's repeated measures, using coefficients from Appendix Table A.10}
\item{Quadratic}{quadratic-trend D variable: the quadratic contrast applied to each participant's repeated measures, using coefficients from Appendix Table A.10}
}
}

\section{Synonym}{

C13T10
}

\examples{
# Load the data
data(chapter_13_table_10)

# Or, alternatively load the data as
data(C13T10)

# View the structure
str(chapter_13_table_10)

# The Linear and Quadratic D variables are formed from the raw three-level
# data of Table 13.2 (chapter_13_table_2) using the orthogonal-polynomial
# coefficients for three equally spaced levels given in Appendix Table A.10:
#   Linear,    coefficients (-1, 0, 1):  Time3 - Time1
#   Quadratic, coefficients (1, -2, 1):  Time1 - 2 * Time2 + Time3
# Because a contrast is unchanged by its sign, Table 13.10 reports the
# quadratic with the equally valid coefficients (-1, 2, -1), that is,
# 2 * Time2 - Time1 - Time3, so that is what is used here.
data(chapter_13_table_2)
derived <- data.frame(
  Linear    = chapter_13_table_2$Time3 - chapter_13_table_2$Time1,
  Quadratic = 2 * chapter_13_table_2$Time2 -
                chapter_13_table_2$Time1 - chapter_13_table_2$Time3
)

# The derived variables reproduce Table 13.10 exactly
all.equal(derived, chapter_13_table_10)

}
\references{
Maxwell, S. E., Delaney, H. D., & Kelley, K. (2027). \emph{Designing experiments and analyzing data:
{A} model comparison perspective} (4th ed.). New York, NY: Routledge.
}
\author{
Ken Kelley \email{kkelley@nd.edu}
}
\keyword{datasets}
