% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{chapter_13_table_12}
\alias{chapter_13_table_12}
\alias{C13T12}
\alias{Chapter_13_Table_12}
\alias{c13t12}
\title{The data used in Chapter 13, Table 12}
\format{
An object of class \code{data.frame} with 8 rows and 2 columns.
}
\source{
\url{https://designingexperiments.com/data/}

Maxwell, S. E., Delaney, H. D., & Kelley, K. (2026). \emph{Designing experiments and
analyzing data: {A} model comparison perspective}. (4th ed.). New York, NY: Routledge.
}
\usage{
data(chapter_13_table_12)
}
\description{
The data used in Chapter 13, Table 12
}
\details{
For the hypothetical data contained in Table 13.2, the linear and quadratic D variables were formed by making use of the appropriate coefficients from Appendix Table A.10. Because the eight participants were measured at three occasions, both a linear and a quadratic effect can be tested. The question of interest in this instance is: "is there a linear and/or quadratic trend exhibited by the group over time?" Recall that in the book (pages 646-647) it was shown that the D variables for linear and quadratic effects led to an omnibus F test of 19.148, which was a value previously obtained for the omnibus effect. Because the particular values chosen for the D variables do not matter (unless it leads to a linear combination of columns), we will focus only on the tests of the individual contrasts when analyzing the data given in Table 12. Because columns one and two already represent the linear and quadratic effect respectively, all that needs to be done is to test mean of the column in order to determine if it differs from zero.
}
\section{Variables}{

\describe{
\item{months30}{hypothetical McCarthy IQ for 30-month-old individuals}
\item{months36}{hypothetical McCarthy IQ for 36-month-old individuals}
\item{months42}{hypothetical McCarthy IQ for 42-month-old individuals}
\item{months48}{hypothetical McCarthy IQ for 48-month-old individuals}
}
}

\section{Synonym}{

C13T12
}

\examples{
# Load the data
data(chapter_13_table_12)

# Or, alternatively load the data as
data(C13T12)

# View the structure
str(chapter_13_table_12)

}
\references{
Maxwell, S. E., Delaney, H. D., & Kelley, K. (2026). \emph{Designing experiments and analyzing data:
{A} model comparison perspective} (4th ed.). New York, NY: Routledge.
}
\author{
Ken Kelley \email{kkelley@nd.edu}
}
\keyword{datasets}
