% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{chapter_13_table_14}
\alias{chapter_13_table_14}
\alias{C13T14}
\alias{Chapter_13_Table_14}
\alias{c13t14}
\title{The data used in Chapter 13, Table 14}
\format{
An object of class \code{data.frame} with 8 rows and 2 columns.
}
\source{
\url{https://designingexperiments.com/data/}

Maxwell, S. E., Delaney, H. D., & Kelley, K. (2018). \emph{Designing experiments and
analyzing data: {A} model comparison perspective}. (3rd ed.). New York, NY: Routledge.
}
\usage{
data(chapter_13_table_14)
}
\description{
The data used in Chapter 13, Table 14
}
\details{
Table 13.14 shows the slope of the least-squares regression line for each of the eight subjects, as well as the score on the linear D variable, reproduced from Table 13.14. There is a strikin relationship between the numbers in the two columns of Table 13.14. Every subject's score on D is 24 times his or her slope.
}
\section{Variables}{

\describe{
\item{slope}{slope of the least-squares regression line for data in Table 13.2}
\item{linear}{Linear D variable for data in Table 13.2}
}
}

\section{Synonym}{

C13T14
}

\examples{
# Load the data
data(chapter_13_table_14)

# Or, alternatively load the data as
data(C13T14)

# View the structure
str(chapter_13_table_14)

}
\references{
Maxwell, S. E., Delaney, H. D., \& Kelley, K. (2018). \emph{Designing experiments and analyzing data:
{A} model comparison perspective} (3rd ed.). New York, NY: Routledge.
}
\author{
Ken Kelley \email{kkelley@nd.edu}
}
\keyword{datasets}
