% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{chapter_13_table_2}
\alias{chapter_13_table_2}
\alias{C13T2}
\alias{Chapter_13_Table_2}
\alias{c13t2}
\title{The data used in Chapter 13, Table 2}
\format{
An object of class \code{data.frame} with 8 rows and 3 columns.
}
\source{
\url{https://designingexperiments.com/data/}

Maxwell, S. E., Delaney, H. D., & Kelley, K. (2018). \emph{Designing experiments and
analyzing data: {A} model comparison perspective}. (3rd ed.). New York, NY: Routledge.
}
\usage{
data(chapter_13_table_2)
}
\description{
The data used in Chapter 13, Table 2
}
\details{
For the hypothetical data contained in Table 13.2, eight participants were measured at three occasions. The question of interest is: "Is there a population mean difference between across Time 1, Time 2, and Time 3 measurement occasions?" Table 13.2 presents hypothetical data for a three-level design. The null hypothesis to be tested is that the population means of scores at all three time points are equal to each.
}
\section{Variables}{

\describe{
\item{time1}{participant score at time 1}
\item{time2}{participant score at time 2}
\item{time3}{participant score at time 3}
}
}

\section{Synonym}{

C13T2
}

\examples{
# Load the data
data(chapter_13_table_2)

# Or, alternatively load the data as
data(C13T2)

# View the structure
str(chapter_13_table_2)

}
\references{
Maxwell, S. E., Delaney, H. D., \& Kelley, K. (2018). \emph{Designing experiments and analyzing data:
{A} model comparison perspective} (3rd ed.). New York, NY: Routledge.
}
\author{
Ken Kelley \email{kkelley@nd.edu}
}
\keyword{datasets}
