% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{chapter_13_table_6}
\alias{chapter_13_table_6}
\alias{C13T6}
\alias{Chapter_13_Table_6}
\alias{c13t6}
\title{The data used in Chapter 13, Table 6}
\format{
An object of class \code{data.frame} with 12 rows and 4 columns.
}
\source{
\url{https://designingexperiments.com/data/}

Maxwell, S. E., Delaney, H. D., & Kelley, K. (2026). \emph{Designing experiments and
analyzing data: {A} model comparison perspective}. (4th ed.). New York, NY: Routledge.
}
\usage{
data(chapter_13_table_6)
}
\description{
The data used in Chapter 13, Table 6
}
\details{
Table 13.6 presents the hypothetical McCarthy IQ scores for 12 subjects (ages 30, 36, 42, or 48 months). The McCarthy data contained in Table 13.6, which was previously analyzed in Table 11.5, is now analyzed using the multivariate approach to repeated measures.
}
\section{Variables}{

\describe{
\item{Months30}{hypothetical McCarthy IQ for 30-month-old individuals}
\item{Months36}{hypothetical McCarthy IQ for 36-month-old individuals}
\item{Months42}{hypothetical McCarthy IQ for 42-month-old individuals}
\item{Months48}{hypothetical McCarthy IQ for 48-month-old individuals}
}
}

\section{Synonym}{

C13T6
}

\examples{
# Load the data
data(chapter_13_table_6)

# Or, alternatively load the data as
data(C13T6)

# View the structure
str(chapter_13_table_6)

}
\references{
Maxwell, S. E., Delaney, H. D., & Kelley, K. (2026). \emph{Designing experiments and analyzing data:
{A} model comparison perspective} (4th ed.). New York, NY: Routledge.
}
\author{
Ken Kelley \email{kkelley@nd.edu}
}
\keyword{datasets}
