% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{chapter_14_exercise_10}
\alias{chapter_14_exercise_10}
\alias{Chapter_14_Exercise_10}
\alias{C14E10}
\alias{c14e10}
\title{The data used in Chapter 14, Exercise 10}
\format{
An object of class \code{data.frame} with 10 rows and 4 columns.
}
\source{
\url{https://designingexperiments.com/data/}

Maxwell, S. E., Delaney, H. D., & Kelley, K. (2027). \emph{Designing experiments and
analyzing data: {A} model comparison perspective}. (4th ed.). New York, NY: Routledge.
}
\usage{
data(chapter_14_exercise_10)
}
\description{
Data from Chapter 14 Exercise 10 of \emph{Designing Experiments and Analyzing Data: A Model Comparison Perspective} (4th edition; Maxwell, Delaney, & Kelley).
}
\details{
\itemize{
\item GridLeft.
\item GridRight.
\item BraceLeft.
\item BraceRight.}
}
\section{Synonym}{

C14E10
}

\examples{
# Load the data
data(chapter_14_exercise_10)

# Or, alternatively load the data as
data(C14E10)

# View the structure
str(chapter_14_exercise_10)

# Brief summary of the data.
summary(chapter_14_exercise_10)

}
\references{
Maxwell, S. E., Delaney, H. D., & Kelley, K. (2027). \emph{Designing experiments and analyzing data:
{A} model comparison perspective} (4th ed.). New York, NY: Routledge.
}
\author{
Ken Kelley \email{kkelley@nd.edu}
}
\keyword{datasets}
