% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{chapter_14_exercise_15}
\alias{chapter_14_exercise_15}
\alias{Chapter_14_Exercise_15}
\alias{C14E15}
\alias{c14e15}
\title{The data used in Chapter 14, Exercise 15}
\format{
An object of class \code{data.frame} with 10 rows and 3 columns.
}
\source{
\url{https://designingexperiments.com/data/}

Maxwell, S. E., Delaney, H. D., & Kelley, K. (2018). \emph{Designing experiments and
analyzing data: {A} model comparison perspective}. (3rd ed.). New York, NY: Routledge.
}
\usage{
data(chapter_14_exercise_15)
}
\description{
Data from Chapter 14 Exercise 15 of \emph{Designing Experiments and Analyzing Data: A Model Comparison Perspective} (3rd edition; Maxwell, Delaney, & Kelley).
}
\details{
\itemize{
\item Baseline.
\item Feedback.
\item Group.}
}
\section{Synonym}{

C14E15
}

\examples{
# Load the data
data(chapter_14_exercise_15)

# Or, alternatively load the data as
data(C14E15)

# View the structure
str(chapter_14_exercise_15)

# Brief summary of the data.
summary(chapter_14_exercise_15)

}
\references{
Maxwell, S. E., Delaney, H. D., & Kelley, K. (2018). \emph{Designing experiments and analyzing data:
{A} model comparison perspective} (3rd ed.). New York, NY: Routledge.
}
\author{
Ken Kelley \email{kkelley@nd.edu}
}
\keyword{datasets}
