% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{chapter_14_exercise_22}
\alias{chapter_14_exercise_22}
\alias{Chapter_14_Exercise_22}
\alias{C14E22}
\alias{c14e22}
\title{The data used in Chapter 14, Exercise 22}
\format{
An object of class \code{data.frame} with 47 rows and 6 columns.
}
\source{
\url{https://designingexperiments.com/data/}

Maxwell, S. E., Delaney, H. D., & Kelley, K. (2018). \emph{Designing experiments and
analyzing data: {A} model comparison perspective}. (3rd ed.). New York, NY: Routledge.
}
\usage{
data(chapter_14_exercise_22)
}
\description{
Data from Chapter 14 Exercise 22 of \emph{Designing Experiments and Analyzing Data: A Model Comparison Perspective} (3rd edition; Maxwell, Delaney, & Kelley).
}
\details{
\itemize{
\item September.
\item November.
\item April.
\item June.
\item July.
\item Group.}
}
\section{Synonym}{

C14E22
}

\examples{
# Load the data
data(chapter_14_exercise_22)

# Or, alternatively load the data as
data(C14E22)

# View the structure
str(chapter_14_exercise_22)

# Brief summary of the data.
summary(chapter_14_exercise_22)

}
\references{
Maxwell, S. E., Delaney, H. D., & Kelley, K. (2018). \emph{Designing experiments and analyzing data:
{A} model comparison perspective} (3rd ed.). New York, NY: Routledge.
}
\author{
Ken Kelley \email{kkelley@nd.edu}
}
\keyword{datasets}
