% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{chapter_14_table_4}
\alias{chapter_14_table_4}
\alias{C14T4}
\alias{Chapter_14_Table_4}
\alias{c14t4}
\title{The data used in Chapter 14, Table 4}
\format{
An object of class \code{data.frame} with 10 rows and 6 columns.
}
\source{
\url{https://designingexperiments.com/data/}

Maxwell, S. E., Delaney, H. D., & Kelley, K. (2018). \emph{Designing experiments and
analyzing data: {A} model comparison perspective}. (3rd ed.). New York, NY: Routledge.
}
\usage{
data(chapter_14_table_4)
}
\description{
The data used in Chapter 14, Table 4
}
\details{
Suppose a perceptual psychologist studying the visual system was interested in determining the extent to which interfering visual stimuli slow the ability to recognize letters. Participants are brought into a laboratory where they are seated in front of a tachistoscope. Variations in the presentations of letters is examined with interest being on the reaction time for target letters presented either in the center of the screen or off centered with and without "noise" accompanying the target letters.
}
\section{Variables}{

\describe{
   \item{abs0}{participant reaction time without noise and with angle 0}
   \item{abs4}{participant reaction time without noise and with angle 4}
   \item{abs8}{participant reaction time without noise and with angle 8}
   \item{pres0}{participant reaction time with noise and with angle 0}
   \item{pres4}{participant reaction time with noise and with angle 4}
   \item{pres8}{participant reaction time with noise and with angle 8}
}
}

\section{Synonym}{

C14T4
}

\examples{
# Load the data
data(chapter_14_table_4)

# Or, alternatively load the data as
data(C14T4)

# View the structure
str(chapter_14_table_4)

}
\references{
Maxwell, S. E., Delaney, H. D., \& Kelley, K. (2018). \emph{Designing experiments and analyzing data:
{A} model comparison perspective} (3rd ed.). New York, NY: Routledge.
}
\author{
Ken Kelley \email{kkelley@nd.edu}
}
\keyword{datasets}
