% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{chapter_14_table_5}
\alias{chapter_14_table_5}
\alias{C14T5}
\alias{Chapter_14_Table_5}
\alias{c14t5}
\title{The data used in Chapter 14, Table 5}
\format{
An object of class \code{data.frame} with 10 rows and 5 columns.
}
\source{
\url{https://designingexperiments.com/data/}

Maxwell, S. E., Delaney, H. D., & Kelley, K. (2018). \emph{Designing experiments and
analyzing data: {A} model comparison perspective}. (3rd ed.). New York, NY: Routledge.
}
\usage{
data(chapter_14_table_5)
}
\description{
The data used in Chapter 14, Table 5
}
\details{
In terms of symbols, let D(1i) represent the linear trend for a given angle. For the hypothetical data contained in Table 14.4, Table 14.5 gives an appropriate and substantively interesting set of D variables. The D variables (rather than the raw data itself) is used because of the benefits and flexibility gained from analyzing the D variables directly (rather than indirectly as we did with the Table 14.4 data).
}
\section{Variables}{

\describe{
\item{d1}{participant D1 variable}
\item{d2}{participant D2 variable}
\item{d3}{participant D3 variable}
\item{d4}{participant D4 variable}
\item{d5}{participant D5 variable}
}
}

\section{Synonym}{

C14T5
}

\examples{
# Load the data
data(chapter_14_table_5)

# Or, alternatively load the data as
data(C14T5)

# View the structure
str(chapter_14_table_5)

}
\references{
Maxwell, S. E., Delaney, H. D., \& Kelley, K. (2018). \emph{Designing experiments and analyzing data:
{A} model comparison perspective} (3rd ed.). New York, NY: Routledge.
}
\author{
Ken Kelley \email{kkelley@nd.edu}
}
\keyword{datasets}
