% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{chapter_14_table_9}
\alias{chapter_14_table_9}
\alias{C14T9}
\alias{Chapter_14_Table_9}
\alias{c14t9}
\title{The data used in Chapter 14, Table 9}
\format{
An object of class \code{data.frame} with 20 rows and 4 columns.
}
\source{
\url{https://designingexperiments.com/data/}

Maxwell, S. E., Delaney, H. D., & Kelley, K. (2026). \emph{Designing experiments and
analyzing data: {A} model comparison perspective}. (4th ed.). New York, NY: Routledge.
}
\usage{
data(chapter_14_table_9)
}
\description{
The data used in Chapter 14, Table 9
}
\details{
These data are identical to those analyzed in chapter 12 (see Tables 12.7 and 12.15) to facilitate comparisons of the multivariate approach and the mixed-model approach.The hypothetical data contained in Table 14.9 contains an additional level of angle (four degrees) that was not considered in Table 14.7.
}
\note{
Renumbered for the 4th edition: in the 3rd edition (AMCP 1.x) these reaction-time data were Table 14.10 (\code{chapter_14_table_10} / \code{C14T10}). The data are unchanged, but note that \code{chapter_14_table_10} now holds different (transformed-score) data.
}
\section{Variables}{

\describe{
\item{Angle0}{participant reaction time at angle 0}
\item{Angle4}{participant reaction time at angle 4}
\item{Angle8}{participant reaction time at angle 8}
\item{Group}{participant age (young or old)}
}
}

\section{Synonym}{

C14T9
}

\examples{
# Load the data
data(chapter_14_table_9)

# Or, alternatively load the data as
data(C14T9)

# View the structure
str(chapter_14_table_9)

}
\references{
Maxwell, S. E., Delaney, H. D., & Kelley, K. (2026). \emph{Designing experiments and analyzing data:
{A} model comparison perspective} (4th ed.). New York, NY: Routledge.
}
\author{
Ken Kelley \email{kkelley@nd.edu}
}
\keyword{datasets}
