% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{chapter_15_exercise_18_univariate}
\alias{chapter_15_exercise_18_univariate}
\alias{Chapter_15_Exercise_18_Univariate}
\alias{C15E18U}
\alias{c15e18u}
\title{The data used in Chapter 15, Exercise 18 (Univariate)}
\format{
An object of class \code{data.frame} with 72 rows and 3 columns.
}
\source{
\url{https://designingexperiments.com/data/}

Maxwell, S. E., Delaney, H. D., & Kelley, K. (2018). \emph{Designing experiments and
analyzing data: {A} model comparison perspective}. (3rd ed.). New York, NY: Routledge.
}
\usage{
data(chapter_15_exercise_18_univariate)
}
\description{
Data from Chapter 15 Exercise 18 (Univariate) of \emph{Designing Experiments and Analyzing Data: A Model Comparison Perspective} (3rd edition; Maxwell, Delaney, & Kelley).
}
\details{
\itemize{
\item ID.
\item Time.
\item Y.}
}
\section{Synonym}{

C15E18U
}

\examples{
# Load the data
data(chapter_15_exercise_18_univariate)

# Or, alternatively load the data as
data(C15E18U)

# View the structure
str(chapter_15_exercise_18_univariate)

# Brief summary of the data.
summary(chapter_15_exercise_18_univariate)

}
\references{
Maxwell, S. E., Delaney, H. D., & Kelley, K. (2018). \emph{Designing experiments and analyzing data:
{A} model comparison perspective} (3rd ed.). New York, NY: Routledge.
}
\author{
Ken Kelley \email{kkelley@nd.edu}
}
\keyword{datasets}
