% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{chapter_15_exercise_19}
\alias{chapter_15_exercise_19}
\alias{Chapter_15_Exercise_19}
\alias{C15E19}
\alias{c15e19}
\title{The data used in Chapter 15, Exercise 19}
\format{
An object of class \code{data.frame} with 24 rows and 4 columns.
}
\source{
\url{https://designingexperiments.com/data/}

Maxwell, S. E., Delaney, H. D., & Kelley, K. (2018). \emph{Designing experiments and
analyzing data: {A} model comparison perspective}. (3rd ed.). New York, NY: Routledge.
}
\usage{
data(chapter_15_exercise_19)
}
\description{
Data from Chapter 15 Exercise 19 of \emph{Designing Experiments and Analyzing Data: A Model Comparison Perspective} (3rd edition; Maxwell, Delaney, & Kelley).
}
\details{
\itemize{
\item Subject.
\item Cognitive70.
\item Cognitive73.
\item Cognitive74.}
}
\section{Synonym}{

C15E19
}

\examples{
# Load the data
data(chapter_15_exercise_19)

# Or, alternatively load the data as
data(C15E19)

# View the structure
str(chapter_15_exercise_19)

# Brief summary of the data.
summary(chapter_15_exercise_19)

}
\references{
Maxwell, S. E., Delaney, H. D., & Kelley, K. (2018). \emph{Designing experiments and analyzing data:
{A} model comparison perspective} (3rd ed.). New York, NY: Routledge.
}
\author{
Ken Kelley \email{kkelley@nd.edu}
}
\keyword{datasets}
