% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{chapter_15_table_1}
\alias{chapter_15_table_1}
\alias{C15T1}
\alias{Chapter_15_Table_1}
\alias{c15t1}
\title{The data used in Chapter 15, Table 1}
\format{
An object of class \code{data.frame} with 12 rows and 4 columns.
}
\source{
\url{https://designingexperiments.com/data/}

Maxwell, S. E., Delaney, H. D., & Kelley, K. (2018). \emph{Designing experiments and
analyzing data: {A} model comparison perspective}. (3rd ed.). New York, NY: Routledge.
}
\usage{
data(chapter_15_table_1)
}
\description{
The data used in Chapter 15, Table 1
}
\details{
The first table in Chapter 15 presents the Hypothetical McCarthy data again (originally shown in Table 11.5). This data set is used throughout the chapter as the to illustrate the discussion given on the mixed model. At the time of this writing the mixed model procedure is SPSS is not as fully developed the SAS analog PROC MIXED. For this reason no analyses using SPSS are provided to replicate the analyses given in Chapter 15.
}
\section{Variables}{

\describe{
\item{months30}{McCarthy IQ score for 30-month-old}
\item{months36}{McCarthy IQ score for 36-month-old}
\item{months42}{McCarthy IQ score for 42-month-old}
\item{months48}{McCarthy IQ score for 48-month-old}}
}

\section{Synonym}{

C15T1
}

\examples{
# Load the data
data(chapter_15_table_1)

# Or, alternatively load the data as
data(C15T1)

# View the structure
str(chapter_15_table_1)

}
\references{
Maxwell, S. E., Delaney, H. D., \& Kelley, K. (2018). \emph{Designing experiments and analyzing data:
{A} model comparison perspective} (3rd ed.). New York, NY: Routledge.
}
\author{
Ken Kelley \email{kkelley@nd.edu}
}
\keyword{datasets}
