% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{chapter_16_table_1}
\alias{chapter_16_table_1}
\alias{C16T1}
\alias{Chapter_16_Table_1}
\alias{c16t1}
\title{The data used in Chapter 16, Table 1}
\format{
An object of class \code{data.frame} with 24 rows and 3 columns.
}
\source{
\url{https://designingexperiments.com/data/}

Maxwell, S. E., Delaney, H. D., & Kelley, K. (2018). \emph{Designing experiments and
analyzing data: {A} model comparison perspective}. (3rd ed.). New York, NY: Routledge.
}
\usage{
data(chapter_16_table_1)
}
\description{
The data used in Chapter 16, Table 1
}
\details{
The first table in Chapter 16 presents the Severity Ratings by Clinical Trainees, which was originally given in Table 10.9. The data set is analyzed again using the multilevel model approach and the results are compared with those obtained in Chapter 10. As a brief background, the goal of the study here is to examine the extent to which female and male clinical psychology graduate student trainees may assign different severity ratings to clients at initial intake.Three female and three male graduate students are randomly selected to participate and each is randomly assigned four clients with whom to do an intake interview, after which each clinical trainee assigns a severity rating to each client, producing the data shown in Table 16.1.
}
\section{Variables}{

\describe{
\item{severity}{severity rating assigned to client by trainee}
\item{trainee}{trainee}
\item{gender}{gender of trainee}}
}

\section{Synonym}{

C16T1
}

\examples{
# Load the data
data(chapter_16_table_1)

# Or, alternatively load the data as
data(C16T1)

# View the structure
str(chapter_16_table_1)

}
\references{
Maxwell, S. E., Delaney, H. D., \& Kelley, K. (2018). \emph{Designing experiments and analyzing data:
{A} model comparison perspective} (3rd ed.). New York, NY: Routledge.
}
\author{
Ken Kelley \email{kkelley@nd.edu}
}
\keyword{datasets}
