% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{chapter_3_exercise_19}
\alias{chapter_3_exercise_19}
\alias{Chapter_3_Exercise_19}
\alias{C3E19}
\alias{c3e19}
\title{The data used in Chapter 3, Exercise 19}
\format{
An object of class \code{data.frame} with 155 rows and 3 columns.
}
\source{
\url{https://designingexperiments.com/data/}

Maxwell, S. E., Delaney, H. D., & Kelley, K. (2018). \emph{Designing experiments and
analyzing data: {A} model comparison perspective}. (3rd ed.). New York, NY: Routledge.
}
\usage{
data(chapter_3_exercise_19)
}
\description{
Data from Chapter 3 Exercise 19 of \emph{Designing Experiments and Analyzing Data: A Model Comparison Perspective} (3rd edition; Maxwell, Delaney, & Kelley).
}
\details{
\itemize{
\item ID.
\item Condition.
\item Anger.}
}
\section{Synonym}{

C3E19
}

\examples{
# Load the data
data(chapter_3_exercise_19)

# Or, alternatively load the data as
data(C3E19)

# View the structure
str(chapter_3_exercise_19)

# Brief summary of the data.
summary(chapter_3_exercise_19)

}
\references{
Maxwell, S. E., Delaney, H. D., & Kelley, K. (2018). \emph{Designing experiments and analyzing data:
{A} model comparison perspective} (3rd ed.). New York, NY: Routledge.
}
\author{
Ken Kelley \email{kkelley@nd.edu}
}
\keyword{datasets}
