% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{chapter_3_exercise_22}
\alias{chapter_3_exercise_22}
\alias{Chapter_3_Exercise_22}
\alias{C3E22}
\alias{c3e22}
\title{The data used in Chapter 3, Exercise 22}
\format{
An object of class \code{data.frame} with 310 rows and 5 columns.
}
\source{
\url{https://designingexperiments.com/data/}

Maxwell, S. E., Delaney, H. D., & Kelley, K. (2018). \emph{Designing experiments and
analyzing data: {A} model comparison perspective}. (3rd ed.). New York, NY: Routledge.
}
\usage{
data(chapter_3_exercise_22)
}
\description{
Data from Chapter 3 Exercise 22 of \emph{Designing Experiments and Analyzing Data: A Model Comparison Perspective} (3rd edition; Maxwell, Delaney, & Kelley).
}
\details{
\itemize{
\item Grade.
\item Treatment.
\item IQPre.
\item IQ4.
\item IQ8.}
}
\section{Synonym}{

C3E22
}

\examples{
# Load the data
data(chapter_3_exercise_22)

# Or, alternatively load the data as
data(C3E22)

# View the structure
str(chapter_3_exercise_22)

# Brief summary of the data.
summary(chapter_3_exercise_22)

}
\references{
Maxwell, S. E., Delaney, H. D., & Kelley, K. (2018). \emph{Designing experiments and analyzing data:
{A} model comparison perspective} (3rd ed.). New York, NY: Routledge.
}
\author{
Ken Kelley \email{kkelley@nd.edu}
}
\keyword{datasets}
