% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{chapter_3_exercise_9}
\alias{chapter_3_exercise_9}
\alias{Chapter_3_Exercise_9}
\alias{C3E9}
\alias{c3e9}
\title{The data used in Chapter 3, Exercise 9}
\format{
An object of class \code{data.frame} with 12 rows and 2 columns.
}
\source{
\url{https://designingexperiments.com/data/}

Maxwell, S. E., Delaney, H. D., & Kelley, K. (2018). \emph{Designing experiments and
analyzing data: {A} model comparison perspective}. (3rd ed.). New York, NY: Routledge.
}
\usage{
data(chapter_3_exercise_9)
}
\description{
Data from Chapter 3 Exercise 9 of \emph{Designing Experiments and Analyzing Data: A Model Comparison Perspective} (3rd edition; Maxwell, Delaney, & Kelley).
}
\details{
\itemize{
\item Group.
\item Scores.}
}
\section{Synonym}{

C3E9
}

\examples{
# Load the data
data(chapter_3_exercise_9)

# Or, alternatively load the data as
data(C3E9)

# View the structure
str(chapter_3_exercise_9)

# Brief summary of the data.
summary(chapter_3_exercise_9)

}
\references{
Maxwell, S. E., Delaney, H. D., \& Kelley, K. (2018). \emph{Designing experiments and analyzing data:
{A} model comparison perspective} (3rd ed.). New York, NY: Routledge.
}
\author{
Ken Kelley \email{kkelley@nd.edu}
}
\keyword{datasets}
