% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{chapter_4_exercise_11}
\alias{chapter_4_exercise_11}
\alias{Chapter_4_Exercise_11}
\alias{C4E11}
\alias{c4e11}
\title{The data used in Chapter 4, Exercise 11}
\format{
An object of class \code{data.frame} with 24 rows and 2 columns.
}
\source{
\url{https://designingexperiments.com/data/}

Maxwell, S. E., Delaney, H. D., & Kelley, K. (2018). \emph{Designing experiments and
analyzing data: {A} model comparison perspective}. (3rd ed.). New York, NY: Routledge.
}
\usage{
data(chapter_4_exercise_11)
}
\description{
Data from Chapter 4 Exercise 11 of \emph{Designing Experiments and Analyzing Data: A Model Comparison Perspective} (3rd edition; Maxwell, Delaney, & Kelley).
}
\details{
\itemize{
\item dv.
\item cond.}
}
\section{Synonym}{

C4E11
}

\examples{
# Load the data
data(chapter_4_exercise_11)

# Or, alternatively load the data as
data(C4E11)

# View the structure
str(chapter_4_exercise_11)

# Brief summary of the data.
summary(chapter_4_exercise_11)

}
\references{
Maxwell, S. E., Delaney, H. D., & Kelley, K. (2018). \emph{Designing experiments and analyzing data:
{A} model comparison perspective} (3rd ed.). New York, NY: Routledge.
}
\author{
Ken Kelley \email{kkelley@nd.edu}
}
\keyword{datasets}
