% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{chapter_4_exercise_18}
\alias{chapter_4_exercise_18}
\alias{Chapter_4_Exercise_18}
\alias{C4E18}
\alias{c4e18}
\title{The data used in Chapter 4, Exercise 18}
\format{
An object of class \code{data.frame} with 12 rows and 2 columns.
}
\source{
\url{https://designingexperiments.com/data/}

Maxwell, S. E., Delaney, H. D., & Kelley, K. (2027). \emph{Designing experiments and
analyzing data: {A} model comparison perspective}. (4th ed.). New York, NY: Routledge.
}
\usage{
data(chapter_4_exercise_18)
}
\description{
Data from Chapter 4 Exercise 18 of \emph{Designing Experiments and Analyzing Data: A Model Comparison Perspective} (4th edition; Maxwell, Delaney, & Kelley). Posttest scores on a fear scale for four treatments for agoraphobia, with three subjects randomly assigned to each therapy; higher scores indicate more severe phobia.
}
\section{Variables}{

\describe{
\item{dv}{posttest score on a fear scale (higher scores indicate more severe phobia)}
\item{cond}{treatment: 1 = rational-emotive (R-E), 2 = psychoanalytic (P), 3 = client-centered (C-C), 4 = behavioral (B)}
}
}

\section{Synonym}{

C4E18
}

\examples{
# Load the data
data(chapter_4_exercise_18)

# Or, alternatively load the data as
data(C4E18)

# View the structure
str(chapter_4_exercise_18)

# Brief summary of the data.
summary(chapter_4_exercise_18)

}
\references{
Maxwell, S. E., Delaney, H. D., & Kelley, K. (2027). \emph{Designing experiments and analyzing data:
{A} model comparison perspective} (4th ed.). New York, NY: Routledge.
}
\author{
Ken Kelley \email{kkelley@nd.edu}
}
\keyword{datasets}
