% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{chapter_4_exercise_21}
\alias{chapter_4_exercise_21}
\alias{Chapter_4_Exercise_21}
\alias{C4E21}
\alias{c4e21}
\title{The data used in Chapter 4, Exercise 21}
\format{
An object of class \code{data.frame} with 39 rows and 2 columns.
}
\source{
\url{https://designingexperiments.com/data/}

Maxwell, S. E., Delaney, H. D., & Kelley, K. (2026). \emph{Designing experiments and
analyzing data: {A} model comparison perspective}. (4th ed.). New York, NY: Routledge.
}
\usage{
data(chapter_4_exercise_21)
}
\description{
Data from Chapter 4 Exercise 21 of \emph{Designing Experiments and Analyzing Data: A Model Comparison Perspective} (4th edition; Maxwell, Delaney, & Kelley). Scores on a multiple-choice memory test from the Kroes and colleagues (2014) study of electroconvulsive therapy (ECT) and memory reconsolidation; higher scores reflect more accurate memory. Participants were randomly assigned to three groups.
}
\section{Variables}{

\describe{
\item{dv}{memory-test score (higher scores reflect more accurate memory)}
\item{cond}{group: 1 = Group A (ECT, tested 24 hours later), 2 = Group B (ECT, tested immediately after the procedure), 3 = Group C (control, no ECT, tested 24 hours later)}
}
}

\section{Synonym}{

C4E21
}

\examples{
# Load the data
data(chapter_4_exercise_21)

# Or, alternatively load the data as
data(C4E21)

# View the structure
str(chapter_4_exercise_21)

# Brief summary of the data.
summary(chapter_4_exercise_21)

}
\references{
Maxwell, S. E., Delaney, H. D., & Kelley, K. (2026). \emph{Designing experiments and analyzing data:
{A} model comparison perspective} (4th ed.). New York, NY: Routledge.
}
\author{
Ken Kelley \email{kkelley@nd.edu}
}
\keyword{datasets}
