% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{chapter_4_table_7}
\alias{chapter_4_table_7}
\alias{C4T7}
\alias{Chapter_4_Table_7}
\alias{c4t7}
\title{The data used in Chapter 4, Table 7}
\format{
An object of class \code{data.frame} with 15 rows and 2 columns.
}
\source{
\url{https://designingexperiments.com/data/}

Maxwell, S. E., Delaney, H. D., & Kelley, K. (2018). \emph{Designing experiments and
analyzing data: {A} model comparison perspective}. (3rd ed.). New York, NY: Routledge.
}
\usage{
data(chapter_4_table_7)
}
\description{
The data used in Chapter 4, Table 7
}
\details{
Data used to demonstrate another property of orthogonal contrasts. The demonstrated principle is that the sums of squares of nonorthogonal contrasts are not additive, yet the sums of squares of orthogonal contrasts can be added to determine the magnitude of the sum of squares they jointly account for.
}
\section{Variables}{

\describe{
\item{dv}{dependent variable}
\item{group}{Group number: 1, 2, or 3}}
}

\section{Synonym}{

C4T7
}

\examples{
# Load the data
data(chapter_4_table_7)

# Or, alternatively load the data as
data(C4T7)

# View the structure
str(chapter_4_table_7)

}
\references{
Maxwell, S. E., Delaney, H. D., \& Kelley, K. (2018). \emph{Designing experiments and analyzing data:
{A} model comparison perspective} (3rd ed.). New York, NY: Routledge.
}
\author{
Ken Kelley \email{kkelley@nd.edu}
}
\keyword{datasets}
