% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{chapter_6_exercise_10}
\alias{chapter_6_exercise_10}
\alias{Chapter_6_Exercise_10}
\alias{C6E10}
\alias{c6e10}
\title{The data used in Chapter 6, Exercise 10}
\format{
An object of class \code{data.frame} with 45 rows and 2 columns.
}
\source{
\url{https://designingexperiments.com/data/}

Maxwell, S. E., Delaney, H. D., & Kelley, K. (2018). \emph{Designing experiments and
analyzing data: {A} model comparison perspective}. (3rd ed.). New York, NY: Routledge.
}
\usage{
data(chapter_6_exercise_10)
}
\description{
Data from Chapter 6 Exercise 10 of \emph{Designing Experiments and Analyzing Data: A Model Comparison Perspective} (3rd edition; Maxwell, Delaney, & Kelley).
}
\details{
\itemize{
\item cond.
\item score.}
}
\section{Synonym}{

C6E10
}

\examples{
# Load the data
data(chapter_6_exercise_10)

# Or, alternatively load the data as
data(C6E10)

# View the structure
str(chapter_6_exercise_10)

# Brief summary of the data.
summary(chapter_6_exercise_10)

}
\references{
Maxwell, S. E., Delaney, H. D., & Kelley, K. (2018). \emph{Designing experiments and analyzing data:
{A} model comparison perspective} (3rd ed.). New York, NY: Routledge.
}
\author{
Ken Kelley \email{kkelley@nd.edu}
}
\keyword{datasets}
