% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{chapter_6_exercise_11}
\alias{chapter_6_exercise_11}
\alias{Chapter_6_Exercise_11}
\alias{C6E11}
\alias{c6e11}
\title{The data used in Chapter 6, Exercise 11}
\format{
An object of class \code{data.frame} with 20 rows and 2 columns.
}
\source{
\url{https://designingexperiments.com/data/}

Maxwell, S. E., Delaney, H. D., & Kelley, K. (2026). \emph{Designing experiments and
analyzing data: {A} model comparison perspective}. (4th ed.). New York, NY: Routledge.
}
\usage{
data(chapter_6_exercise_11)
}
\description{
Data from Chapter 6 Exercise 11 of \emph{Designing Experiments and Analyzing Data: A Model Comparison Perspective} (4th edition; Maxwell, Delaney, & Kelley). The mean number of incorrect turns made by rats learning a maze after exposure to one of four equally spaced dosage levels of a drug, with five animals per level. Because drug dosage is quantitative, these data are suited to trend analysis.
}
\section{Variables}{

\describe{
\item{Errors}{mean number of incorrect turns made over five trials}
\item{Dosage}{drug dosage level, equally spaced in units of size 1 (1, 2, 3, 4)}
}
}

\section{Synonym}{

C6E11
}

\examples{
# Load the data
data(chapter_6_exercise_11)

# Or, alternatively load the data as
data(C6E11)

# View the structure
str(chapter_6_exercise_11)

# Brief summary of the data.
summary(chapter_6_exercise_11)

}
\references{
Maxwell, S. E., Delaney, H. D., & Kelley, K. (2026). \emph{Designing experiments and analyzing data:
{A} model comparison perspective} (4th ed.). New York, NY: Routledge.
}
\author{
Ken Kelley \email{kkelley@nd.edu}
}
\keyword{datasets}
