% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{chapter_6_exercise_16}
\alias{chapter_6_exercise_16}
\alias{Chapter_6_Exercise_16}
\alias{C6E16}
\alias{c6e16}
\title{The data used in Chapter 6, Exercise 16}
\format{
An object of class \code{data.frame} with 91 rows and 5 columns.
}
\source{
\url{https://designingexperiments.com/data/}

Maxwell, S. E., Delaney, H. D., & Kelley, K. (2018). \emph{Designing experiments and
analyzing data: {A} model comparison perspective}. (3rd ed.). New York, NY: Routledge.
}
\usage{
data(chapter_6_exercise_16)
}
\description{
Data from Chapter 6 Exercise 16 of \emph{Designing Experiments and Analyzing Data: A Model Comparison Perspective} (3rd edition; Maxwell, Delaney, & Kelley).
}
\details{
\itemize{
\item id.
\item group.
\item y.
\item latency.
\item latency_2.}
}
\section{Synonym}{

C6E16
}

\examples{
# Load the data
data(chapter_6_exercise_16)

# Or, alternatively load the data as
data(C6E16)

# View the structure
str(chapter_6_exercise_16)

# Brief summary of the data.
summary(chapter_6_exercise_16)

}
\references{
Maxwell, S. E., Delaney, H. D., & Kelley, K. (2018). \emph{Designing experiments and analyzing data:
{A} model comparison perspective} (3rd ed.). New York, NY: Routledge.
}
\author{
Ken Kelley \email{kkelley@nd.edu}
}
\keyword{datasets}
