% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{chapter_6_table_1}
\alias{chapter_6_table_1}
\alias{C6T1}
\alias{Chapter_6_Table_1}
\alias{c6t1}
\title{The data used in Chapter 6, Table 1}
\format{
An object of class \code{data.frame} with 24 rows and 2 columns.
}
\source{
\url{https://designingexperiments.com/data/}

Maxwell, S. E., Delaney, H. D., & Kelley, K. (2018). \emph{Designing experiments and
analyzing data: {A} model comparison perspective}. (3rd ed.). New York, NY: Routledge.
}
\usage{
data(chapter_6_table_1)
}
\description{
The data used in Chapter 6, Table 1
}
\details{
Recall scores for 24 children who have been randomly assigned to one of four experimental conditions where there are 6 children in each condition. The experimental conditions of interest are 1, 2, 3, and 4 minutes where the number of minutes is the amount of time the child is allotted to study a list of words before attempting to recall the words. The dependent variable (i.e., the recall scores) are the number of words the child is able to recall after a brief interference task. The first hypothesis of interest is whether the number of words recalled is linearly related to the number of minutes spent studying.
}
\section{Variables}{

\describe{
\item{recall}{the number of words recalled by the child after the study time expires}
\item{minutes}{the amount of time, in minutes, the child was permitted to study}}
}

\section{Synonym}{

C6T1
}

\examples{
# Load the data
data(chapter_6_table_1)

# Or, alternatively load the data as
data(C6T1)

# View the structure
str(chapter_6_table_1)

}
\references{
Maxwell, S. E., Delaney, H. D., \& Kelley, K. (2018). \emph{Designing experiments and analyzing data:
{A} model comparison perspective} (3rd ed.). New York, NY: Routledge.
}
\author{
Ken Kelley \email{kkelley@nd.edu}
}
\keyword{datasets}
