% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{chapter_7_exercise_12}
\alias{chapter_7_exercise_12}
\alias{Chapter_7_Exercise_12}
\alias{C7E12}
\alias{c7e12}
\title{The data used in Chapter 7, Exercise 12}
\format{
An object of class \code{data.frame} with 32 rows and 3 columns.
}
\source{
\url{https://designingexperiments.com/data/}

Maxwell, S. E., Delaney, H. D., & Kelley, K. (2018). \emph{Designing experiments and
analyzing data: {A} model comparison perspective}. (3rd ed.). New York, NY: Routledge.
}
\usage{
data(chapter_7_exercise_12)
}
\description{
Data from Chapter 7 Exercise 12 of \emph{Designing Experiments and Analyzing Data: A Model Comparison Perspective} (3rd edition; Maxwell, Delaney, & Kelley).
}
\details{
\itemize{
\item ALevel.
\item BLevel.
\item Score.}
}
\section{Synonym}{

C7E12
}

\examples{
# Load the data
data(chapter_7_exercise_12)

# Or, alternatively load the data as
data(C7E12)

# View the structure
str(chapter_7_exercise_12)

# Brief summary of the data.
summary(chapter_7_exercise_12)

}
\references{
Maxwell, S. E., Delaney, H. D., & Kelley, K. (2018). \emph{Designing experiments and analyzing data:
{A} model comparison perspective} (3rd ed.). New York, NY: Routledge.
}
\author{
Ken Kelley \email{kkelley@nd.edu}
}
\keyword{datasets}
