% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{chapter_7_exercise_18}
\alias{chapter_7_exercise_18}
\alias{Chapter_7_Exercise_18}
\alias{C7E18}
\alias{c7e18}
\title{The data used in Chapter 7, Exercise 18}
\format{
An object of class \code{data.frame} with 22 rows and 3 columns.
}
\source{
\url{https://designingexperiments.com/data/}

Maxwell, S. E., Delaney, H. D., & Kelley, K. (2018). \emph{Designing experiments and
analyzing data: {A} model comparison perspective}. (3rd ed.). New York, NY: Routledge.
}
\usage{
data(chapter_7_exercise_18)
}
\description{
Data from Chapter 7 Exercise 18 of \emph{Designing Experiments and Analyzing Data: A Model Comparison Perspective} (3rd edition; Maxwell, Delaney, & Kelley).
}
\details{
\itemize{
\item level.
\item gender.
\item salary.}
}
\section{Synonym}{

C7E18
}

\examples{
# Load the data
data(chapter_7_exercise_18)

# Or, alternatively load the data as
data(C7E18)

# View the structure
str(chapter_7_exercise_18)

# Brief summary of the data.
summary(chapter_7_exercise_18)

}
\references{
Maxwell, S. E., Delaney, H. D., & Kelley, K. (2018). \emph{Designing experiments and analyzing data:
{A} model comparison perspective} (3rd ed.). New York, NY: Routledge.
}
\author{
Ken Kelley \email{kkelley@nd.edu}
}
\keyword{datasets}
