% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{chapter_7_exercise_21}
\alias{chapter_7_exercise_21}
\alias{Chapter_7_Exercise_21}
\alias{C7E21}
\alias{c7e21}
\title{The data used in Chapter 7, Exercise 21}
\format{
An object of class \code{data.frame} with 40 rows and 3 columns.
}
\source{
\url{https://designingexperiments.com/data/}

Maxwell, S. E., Delaney, H. D., & Kelley, K. (2026). \emph{Designing experiments and
analyzing data: {A} model comparison perspective}. (4th ed.). New York, NY: Routledge.
}
\usage{
data(chapter_7_exercise_21)
}
\description{
Data from Chapter 7 Exercise 21 of \emph{Designing Experiments and Analyzing Data: A Model Comparison Perspective} (4th edition; Maxwell, Delaney, & Kelley). Trust scores in a 2 (gender) by 2 (drug) between-subjects factorial design in which 20 females and 20 males were randomly assigned to receive either oxytocin or a placebo; trust was measured in a variation of the prisoner's dilemma, and higher scores indicate more trust.
}
\section{Variables}{

\describe{
\item{Gender}{1 = female, 2 = male}
\item{Drug}{1 = oxytocin, 2 = placebo}
\item{Trust}{trust score (higher scores indicate more trust)}
}
}

\section{Synonym}{

C7E21
}

\examples{
# Load the data
data(chapter_7_exercise_21)

# Or, alternatively load the data as
data(C7E21)

# View the structure
str(chapter_7_exercise_21)

# Brief summary of the data.
summary(chapter_7_exercise_21)

}
\references{
Maxwell, S. E., Delaney, H. D., & Kelley, K. (2026). \emph{Designing experiments and analyzing data:
{A} model comparison perspective} (4th ed.). New York, NY: Routledge.
}
\author{
Ken Kelley \email{kkelley@nd.edu}
}
\keyword{datasets}
