% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{chapter_7_exercise_24}
\alias{chapter_7_exercise_24}
\alias{Chapter_7_Exercise_24}
\alias{C7E24}
\alias{c7e24}
\title{The data used in Chapter 7, Exercise 24}
\format{
An object of class \code{data.frame} with 56 rows and 4 columns.
}
\source{
\url{https://designingexperiments.com/data/}

Maxwell, S. E., Delaney, H. D., & Kelley, K. (2026). \emph{Designing experiments and
analyzing data: {A} model comparison perspective}. (4th ed.). New York, NY: Routledge.
}
\usage{
data(chapter_7_exercise_24)
}
\description{
Data from Chapter 7 Exercise 24 of \emph{Designing Experiments and Analyzing Data: A Model Comparison Perspective} (4th edition; Maxwell, Delaney, & Kelley).
}
\details{
\itemize{
\item id.
\item thought.
\item complexity.
\item attitude.}
}
\section{Synonym}{

C7E24
}

\examples{
# Load the data
data(chapter_7_exercise_24)

# Or, alternatively load the data as
data(C7E24)

# View the structure
str(chapter_7_exercise_24)

# Brief summary of the data.
summary(chapter_7_exercise_24)

}
\references{
Maxwell, S. E., Delaney, H. D., & Kelley, K. (2026). \emph{Designing experiments and analyzing data:
{A} model comparison perspective} (4th ed.). New York, NY: Routledge.
}
\author{
Ken Kelley \email{kkelley@nd.edu}
}
\keyword{datasets}
