% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{chapter_7_exercise_9}
\alias{chapter_7_exercise_9}
\alias{Chapter_7_Exercise_9}
\alias{C7E9}
\alias{c7e9}
\title{The data used in Chapter 7, Exercise 9}
\format{
An object of class \code{data.frame} with 48 rows and 3 columns.
}
\source{
\url{https://designingexperiments.com/data/}

Maxwell, S. E., Delaney, H. D., & Kelley, K. (2018). \emph{Designing experiments and
analyzing data: {A} model comparison perspective}. (3rd ed.). New York, NY: Routledge.
}
\usage{
data(chapter_7_exercise_9)
}
\description{
Data from Chapter 7 Exercise 9 of \emph{Designing Experiments and Analyzing Data: A Model Comparison Perspective} (3rd edition; Maxwell, Delaney, & Kelley).
}
\details{
\itemize{
\item Treatment.
\item Level.
\item Score.}
}
\section{Synonym}{

C7E9
}

\examples{
# Load the data
data(chapter_7_exercise_9)

# Or, alternatively load the data as
data(C7E9)

# View the structure
str(chapter_7_exercise_9)

# Brief summary of the data.
summary(chapter_7_exercise_9)

}
\references{
Maxwell, S. E., Delaney, H. D., & Kelley, K. (2018). \emph{Designing experiments and analyzing data:
{A} model comparison perspective} (3rd ed.). New York, NY: Routledge.
}
\author{
Ken Kelley \email{kkelley@nd.edu}
}
\keyword{datasets}
