% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{chapter_7_table_11}
\alias{chapter_7_table_11}
\alias{C7T11}
\alias{Chapter_7_Table_11}
\alias{c7t11}
\title{The data used in Chapter 7, Table 11}
\format{
An object of class \code{data.frame} with 45 rows and 3 columns.
}
\source{
\url{https://designingexperiments.com/data/}

Maxwell, S. E., Delaney, H. D., & Kelley, K. (2018). \emph{Designing experiments and
analyzing data: {A} model comparison perspective}. (3rd ed.). New York, NY: Routledge.
}
\usage{
data(chapter_7_table_11)
}
\description{
The data used in Chapter 7, Table 11
}
\details{
Table 7.11 presents this hypothetical data for 15 amnesiacs, 15 Huntington individuals, and 15 controls. The data represents a two way factorial design where diagnosis and task are fully crossed, each with three levels. Of interest for the results displayed in Table 7.12 is whether the interaction contrast specified in Figure 7.3 and 7.4 is statistically significant. Namely the question pertains to whether the relationship of the mean of grammar and classification versus recognition differs for those in the amnesic and Huntington's group. The procedure SPSS syntax procedure MANOVA is very general and can handle many types of analyses. Interaction contrasts are easily performed in this procedure.

Consider an example of a cognitive neuroscience study of patient groups. Specifically, suppose that a certain theory implies that amnesic patients will have a deficit in explicit memory but not in implicit memory. According to this theory, Huntington patients, on the other hand, will be just the opposite: They will have no deficit in explicit memory, but will have a deficit in implicit memory. Further suppose that a study is designed yielding a 3x3 factorial design to test this theory. The rows of this study will represent three types of individuals: amnesic patients, Huntington patients, and a control group of individuals with no known neurological disorder. Each research participant will be randomly assigned to one of three tasks: (1) artificial grammar task, which consists of classifying letter sequences as either following or not following grammatical rules; (2) classification learning task, which consists of classifying hypothetical patients as either having or not having a certain disease based on symptoms probabilistically related to the disease; and (3) recognition memory task, which consists of recognizing particular stimuli as stimuli that have previously been presented during the task.
}
\section{Variables}{

\describe{
\item{score}{blood pressure}
\item{feedback}{the likelihood of there being a biofeedback or drug main effect}
\item{drug}{the level of the drug factor}}
}

\section{Synonym}{

C7T11
}

\examples{
# Load the data
data(chapter_7_table_11)

# Or, alternatively load the data as
data(C7T11)

# View the structure
str(chapter_7_table_11)

}
\references{
Maxwell, S. E., Delaney, H. D., \& Kelley, K. (2018). \emph{Designing experiments and analyzing data:
{A} model comparison perspective} (3rd ed.). New York, NY: Routledge.
}
\author{
Ken Kelley \email{kkelley@nd.edu}
}
\keyword{datasets}
