% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{chapter_7_table_15}
\alias{chapter_7_table_15}
\alias{C7T15}
\alias{Chapter_7_Table_15}
\alias{c7t15}
\title{The data used in Chapter 7, Table 15}
\format{
An object of class \code{data.frame} with 22 rows and 3 columns.
}
\source{
\url{https://designingexperiments.com/data/}

Maxwell, S. E., Delaney, H. D., & Kelley, K. (2026). \emph{Designing experiments and
analyzing data: {A} model comparison perspective}. (4th ed.). New York, NY: Routledge.
}
\usage{
data(chapter_7_table_15)
}
\description{
The data used in Chapter 7, Table 15
}
\details{
The following hypothetical salary data represents a nonorthogonal two-by-two factorial design. The first factor (sex) is crossed with college (degree or no degree). The primary question of interest is whether or not there is sex discrimination in terms of salary.

The data in Table 7.15 presents hypothetical data (in thousands) for 12 females and 10 males who have just been hired by the organization. The mean salary for the 12 females is $22,333, whereas the mean for the 10 males is $22,100. The data in Table 7.15 also contains information about an additional characteristic of employees, namely whether they received a college degree. It is clear from the data that a majority of the new female employees are college graduates, whereas a majority of the males are not.
}
\section{Variables}{

\describe{
\item{sex}{gender (male vs female)}
\item{educ}{education level (degree vs no degree)}
\item{salary}{salary (in thousands)}
}
}

\section{Synonym}{
 
C7T15
}

\examples{
# Load the data
data(chapter_7_table_15)

# Or, alternatively load the data as
data(C7T15)

# View the structure
str(chapter_7_table_15)

}
\references{
Maxwell, S. E., Delaney, H. D., & Kelley, K. (2026). \emph{Designing experiments and analyzing data:
{A} model comparison perspective} (4th ed.). New York, NY: Routledge.
}
\author{
Ken Kelley \email{kkelley@nd.edu}
}
\keyword{datasets}
