% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{chapter_8_exercise_17}
\alias{chapter_8_exercise_17}
\alias{Chapter_8_Exercise_17}
\alias{C8E17}
\alias{c8e17}
\title{The data used in Chapter 8, Exercise 17}
\format{
An object of class \code{data.frame} with 54 rows and 4 columns.
}
\source{
\url{https://designingexperiments.com/data/}

Maxwell, S. E., Delaney, H. D., & Kelley, K. (2018). \emph{Designing experiments and
analyzing data: {A} model comparison perspective}. (3rd ed.). New York, NY: Routledge.
}
\usage{
data(chapter_8_exercise_17)
}
\description{
Data from Chapter 8 Exercise 17 of \emph{Designing Experiments and Analyzing Data: A Model Comparison Perspective} (3rd edition; Maxwell, Delaney, & Kelley).
}
\details{
\itemize{
\item BehavioralAvoidance.
\item Condition.
\item Phobia.
\item Gender.}
}
\section{Synonym}{

C8E17
}

\examples{
# Load the data
data(chapter_8_exercise_17)

# Or, alternatively load the data as
data(C8E17)

# View the structure
str(chapter_8_exercise_17)

# Brief summary of the data.
summary(chapter_8_exercise_17)

}
\references{
Maxwell, S. E., Delaney, H. D., & Kelley, K. (2018). \emph{Designing experiments and analyzing data:
{A} model comparison perspective} (3rd ed.). New York, NY: Routledge.
}
\author{
Ken Kelley \email{kkelley@nd.edu}
}
\keyword{datasets}
