% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{chapter_9_exercise_15}
\alias{chapter_9_exercise_15}
\alias{Chapter_9_Exercise_15}
\alias{C9E15}
\alias{c9e15}
\title{The data used in Chapter 9, Exercise 15}
\format{
An object of class \code{data.frame} with 310 rows and 6 columns.
}
\source{
\url{https://designingexperiments.com/data/}

Maxwell, S. E., Delaney, H. D., & Kelley, K. (2018). \emph{Designing experiments and
analyzing data: {A} model comparison perspective}. (3rd ed.). New York, NY: Routledge.
}
\usage{
data(chapter_9_exercise_15)
}
\description{
Data from Chapter 9 Exercise 15 of \emph{Designing Experiments and Analyzing Data: A Model Comparison Perspective} (3rd edition; Maxwell, Delaney, & Kelley).
}
\details{
\itemize{
\item ID.
\item Condition.
\item EmotClose.
\item Anger.}
}
\section{Synonym}{

C9E15
}

\examples{
# Load the data
data(chapter_9_exercise_15)

# Or, alternatively load the data as
data(C9E15)

# View the structure
str(chapter_9_exercise_15)

# Brief summary of the data.
summary(chapter_9_exercise_15)

}
\references{
Maxwell, S. E., Delaney, H. D., & Kelley, K. (2018). \emph{Designing experiments and analyzing data:
{A} model comparison perspective} (3rd ed.). New York, NY: Routledge.
}
\author{
Ken Kelley \email{kkelley@nd.edu}
}
\keyword{datasets}
