% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{chapter_9_extension_exercise_1}
\alias{chapter_9_extension_exercise_1}
\alias{Chapter_9_Extension_Exercise_1}
\alias{C9ExtE1}
\alias{c9exte1}
\title{The data used in Chapter 9 Extension, Exercise 1}
\format{
An object of class \code{data.frame} with 140 rows and 6 columns.
}
\source{
\url{https://designingexperiments.com/data/}

Maxwell, S. E., Delaney, H. D., & Kelley, K. (2018). \emph{Designing experiments and
analyzing data: {A} model comparison perspective}. (3rd ed.). New York, NY: Routledge.

Sturge-Apple, M. L.,  Suor, J. H., Davies, P. T., Cicchetti, D., Skibo, M. A., \& Rogosch, F. A. (2016). Vagal tone and children's delay of gratification: Differential sensitivity in resource-poor and resource-rich environments. \emph{Psychological Science}, \emph{27}, 885--893.
}
\usage{
data(chapter_9_extension_exercise_1)
}
\description{
Data from Chapter 9 Extension Exercise 1 of \emph{Designing Experiments and Analyzing Data: A Model Comparison Perspective} (3rd edition; Maxwell, Delaney, & Kelley).
}
\details{
\itemize{
\item ID.
\item RSA.
\item Delay.
\item SES_group.
\item MaxDelay.
\item RSAdev.}
}
\section{Synonym}{

C9ExtE1
}

\examples{
# Load the data
data(chapter_9_extension_exercise_1)

# Or, alternatively load the data as
data(C9ExtE1)

# View the structure
str(chapter_9_extension_exercise_1)

# Brief summary of the data.
summary(chapter_9_extension_exercise_1)

}
\references{
Maxwell, S. E., Delaney, H. D., & Kelley, K. (2018). \emph{Designing experiments and analyzing data: {A} model comparison perspective} (3rd ed.). New York, NY: Routledge.
}
\author{
Ken Kelley \email{kkelley@nd.edu}
}
\keyword{datasets}
