% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{chapter_9_extension_exercise_2}
\alias{chapter_9_extension_exercise_2}
\alias{Chapter_9_Extension_Exercise_2}
\alias{C9ExtE2}
\alias{c9exte2}
\title{The data used in Chapter 9 Extension, Exercise 2}
\format{
An object of class \code{data.frame} with 168 rows and 6 columns.
}
\source{
\url{https://designingexperiments.com/data/}

Maxwell, S. E., Delaney, H. D., & Kelley, K. (2026). \emph{Designing experiments and
analyzing data: {A} model comparison perspective}. (4th ed.). New York, NY: Routledge.
}
\usage{
data(chapter_9_extension_exercise_2)
}
\description{
Data from Chapter 9 Extension Exercise 2 of \emph{Designing Experiments and Analyzing Data: A Model Comparison Perspective} (4th edition; Maxwell, Delaney, & Kelley).
}
\details{
\itemize{
\item ParticipantNumber.
\item Group.
\item BaseHrsDrkTypWk.
\item FolHrsDrkTypWk.
\item DiffBaseFolHrsDrk.
\item BaseHrsCtrd.}
}
\section{Synonym}{

C9ExtE2
}

\examples{
# Load the data
data(chapter_9_extension_exercise_2)

# Or, alternatively load the data as
data(C9ExtE2)

# View the structure
str(chapter_9_extension_exercise_2)

# Brief summary of the data.
summary(chapter_9_extension_exercise_2)

}
\references{
Maxwell, S. E., Delaney, H. D., & Kelley, K. (2026). \emph{Designing experiments and analyzing data:
{A} model comparison perspective} (4th ed.). New York, NY: Routledge.
}
\author{
Ken Kelley \email{kkelley@nd.edu}
}
\keyword{datasets}
