% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{chapter_9_extension_figures_4_and_5}
\alias{chapter_9_extension_figures_4_and_5}
\alias{Chapter_9_Extension_Figures_4_and_5}
\alias{c9extfigs4and5}
\alias{C9ExtFigs4and5}
\title{The data used in Chapter 9 Extension Figures 4 and 5}
\format{
An object of class \code{data.frame} with 310 rows and 10 columns.
}
\source{
\url{https://designingexperiments.com/data/}

Maxwell, S. E., Delaney, H. D., & Kelley, K. (2018). \emph{Designing experiments and
analyzing data: {A} model comparison perspective}. (3rd ed.). New York, NY: Routledge.
}
\usage{
data(chapter_9_extension_figures_4_and_5)
}
\description{
Data used in the Chapter 9 Extension, Figures 4 and 5, of \emph{Designing Experiments and Analyzing Data: A Model Comparison Perspective} (3rd edition; Maxwell, Delaney, & Kelley).
}
\details{
\itemize{
\item Grade
\item Treatment
\item IQPre
\item IQ4
\item IQ8
\item AvPost
\item IQPre_Mean
\item IQPre_Centered
\item TxX
\item Constant1}
}
\section{Synonym}{

C9ExtE1
}

\examples{
# Load the data
data(chapter_9_extension_figures_4_and_5)

# Or, alternatively load the data as
data(C9ExtFigs4and5)

# View the structure
str(chapter_9_extension_figures_4_and_5)

# Brief summary of the data.
summary(chapter_9_extension_figures_4_and_5)

}
\references{
Maxwell, S. E., Delaney, H. D., & Kelley, K. (2018). \emph{Designing experiments and analyzing data: {A} model comparison perspective} (3rd ed.). New York, NY: Routledge.
}
\author{
Ken Kelley \email{kkelley@nd.edu}
}
\keyword{datasets}
