% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{chapter_9_extension_table_1}
\alias{chapter_9_extension_table_1}
\alias{C9ExtT1}
\alias{Chapter_9_Extension_Table_1}
\alias{c9extt1}
\title{The data used in Chapter 9, Extension Table 1}
\format{
An object of class \code{data.frame} with 6 rows and 3 columns.
}
\source{
\url{https://designingexperiments.com/data/}

Maxwell, S. E., Delaney, H. D., & Kelley, K. (2018). \emph{Designing experiments and
analyzing data: {A} model comparison perspective}. (3rd ed.). New York, NY: Routledge.
}
\usage{
data(chapter_9_extension_table_1)
}
\description{
The data used in Chapter 9, Extension Table 1
}
\details{
Table 9E.1 shows the data from Table 9.1 after some minor modifications to reflect heterogeneity of regression. The data were altered in such a way that the means in both groups are the same as in the original example, as is the pooled within-group slope.
}
\section{Variables}{

\describe{
\item{group}{the group (treatment group vs wait-list control group)}
\item{x}{the weight lost by the control group}
\item{y}{the weight lost by the treatment group}
}
}

\section{Synonym}{

C9ExtT1
}

\examples{
# Load the data
data(chapter_9_extension_table_1)

# Or, alternatively load the data as
data(C9ExtT1)

# View the structure
str(chapter_9_extension_table_1)

}
\references{
Maxwell, S. E., Delaney, H. D., \& Kelley, K. (2018). \emph{Designing experiments and analyzing data:
{A} model comparison perspective} (3rd ed.). New York, NY: Routledge.
}
\author{
Ken Kelley \email{kkelley@nd.edu}
}
\keyword{datasets}
