% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{chapter_9_table_11}
\alias{chapter_9_table_11}
\alias{C9T11}
\alias{Chapter_9_Table_11}
\alias{c9t11}
\title{The data used in Chapter 9, Table 11}
\format{
An object of class \code{data.frame} with 18 rows and 4 columns.
}
\source{
\url{https://designingexperiments.com/data/}

Maxwell, S. E., Delaney, H. D., & Kelley, K. (2018). \emph{Designing experiments and
analyzing data: {A} model comparison perspective}. (3rd ed.). New York, NY: Routledge.
}
\usage{
data(chapter_9_table_11)
}
\description{
The data used in Chapter 9, Table 11
}
\details{
The question of interest in the present situation assumes that there are three blocks of elderly participants, six per block. The elderly participants are sorted into the three blocks as a function of their age. The purpose of the study was to assess the effect of age on motor control, measured by the number of errors on a certain task (where there were three tasks). The goal here is to replicate the results from the top portion of Table 9.12 (At this point performing the ANOVA and the ANCOVA also included in Table 9.12 should be straightforward.#'
}
\section{Variables}{

\describe{
\item{block}{a numeric vector between 1 and 3, equal to the block of elderly participants (6 per block)}
\item{task}{the task given}
\item{x}{age}
\item{y}{error scores}
}
}

\section{Synonym}{

C9T11
}

\examples{
# Load the data
data(chapter_9_table_11)

# Or, alternatively load the data as
data(C9T11)

# View the structure
str(chapter_9_table_11)

}
\references{
Maxwell, S. E., Delaney, H. D., \& Kelley, K. (2018). \emph{Designing experiments and analyzing data:
{A} model comparison perspective} (3rd ed.). New York, NY: Routledge.
}
\author{
Ken Kelley \email{kkelley@nd.edu}
}
\keyword{datasets}
