% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{tutorial_1_table_1}
\alias{tutorial_1_table_1}
\alias{T1T1}
\alias{t1t1}
\title{The data used in Tutorial 1, Table 1}
\format{
An object of class \code{data.frame} with 102 rows and 1 columns.
}
\source{
\url{https://designingexperiments.com/data/}

Maxwell, S. E., Delaney, H. D., & Kelley, K. (2018). \emph{Designing experiments and
analyzing data: {A} model comparison perspective}. (3rd ed.). New York, NY: Routledge.
}
\usage{
data(tutorial_1_table_1)
}
\description{
Data from Tutorial 1 Table 1 of \emph{Designing Experiments and Analyzing Data: A Model Comparison Perspective} (3rd edition; Maxwell, Delaney, & Kelley).
}
\details{
\itemize{
\item Group.
\item Score.}
}
\section{Synonym}{

T1T1
}

\examples{
# Load the data
data(tutorial_1_table_1)

# Or, alternatively load the data as
data(T1T1)

# View the structure
str(tutorial_1_table_1)

# Brief summary of the data.
summary(tutorial_1_table_1)

}
\references{
Maxwell, S. E., Delaney, H. D., & Kelley, K. (2018). \emph{Designing experiments and analyzing data:
{A} model comparison perspective} (3rd ed.). New York, NY: Routledge.
}
\author{
Ken Kelley \email{kkelley@nd.edu}
}
\keyword{datasets}
