% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{tutorial_2_table_2}
\alias{tutorial_2_table_2}
\alias{T2T2}
\alias{t2t2}
\title{The data used in Tutorial 2, Table 2}
\format{
An object of class \code{data.frame} with 8 rows and 4 columns.
}
\source{
\url{https://designingexperiments.com/data/}

Maxwell, S. E., Delaney, H. D., & Kelley, K. (2026). \emph{Designing experiments and
analyzing data: {A} model comparison perspective}. (4th ed.). New York, NY: Routledge.
}
\usage{
data(tutorial_2_table_2)
}
\description{
Data from Tutorial 2 Table 2 of \emph{Designing Experiments and Analyzing Data: A Model Comparison Perspective} (4th edition; Maxwell, Delaney, & Kelley).
}
\details{
\itemize{
\item Y.
\item X1.
\item X2.
\item X3.}
}
\section{Synonym}{

T2T2
}

\examples{
# Load the data
data(tutorial_2_table_2)

# Or, alternatively load the data as
data(T2T2)

# View the structure
str(tutorial_2_table_2)

# Brief summary of the data.
summary(tutorial_2_table_2)

}
\references{
Maxwell, S. E., Delaney, H. D., & Kelley, K. (2026). \emph{Designing experiments and analyzing data:
{A} model comparison perspective} (4th ed.). New York, NY: Routledge.
}
\author{
Ken Kelley \email{kkelley@nd.edu}
}
\keyword{datasets}
