% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{chapter_10_table_5}
\alias{chapter_10_table_5}
\alias{C10T5}
\alias{Chapter_10_Table_5}
\alias{c10t5}
\title{The data used in Chapter 10, Table 5}
\format{
An object of class \code{data.frame} with 40 rows and 3 columns.
}
\source{
\url{https://designingexperiments.com/data/}

Maxwell, S. E., Delaney, H. D., & Kelley, K. (2018). \emph{Designing experiments and
analyzing data: {A} model comparison perspective}. (3rd ed.). New York, NY: Routledge.
}
\usage{
data(chapter_10_table_5)
}
\description{
The data used in Chapter 10, Table 5
}
\details{
Assume that an educational products firm markets study programs to help high school students prepare for college entrance exams such as the ACT, and wants to compare a new computer-based training program with their standard packet of printed materials. The firm would like to be able to generalize to all American high schools but only has the resources to conduct a study in a few schools. Thus, assume four high schools are selected at random from a listing of all public schools in the country. Volunteers from the junior class at these schools are solicited to take part in an eight-session after-school study program. Ten students from each school are permitted to take part, and equal numbers from each school are assigned randomly to the two study programs. Designating the type of study program as factor A (a1 designates the computer-based program and a2 designates the standard paper-and-pencil program) and the particular school as factor B, assume the data in Table 10.5 are obtained.

The data consists of simulated ACT scores from 40 participants where 10 participants were selected from each of four schools. It is assumed that the schools are randomly selected from a population of schools in America in order to generalize the results found. Two schools (and thus, 20 participants) are randomly assigned to the computer-based ACT training program, while the other two schools are randomly assigned to the standard paper-and-pencil program in order to assess the effectiveness of these different types of programs.

The primary hypothesis of interest is whether the standard paper-and-pencil and computer-based ACT training programs differ in effectiveness.
}
\section{Variables}{

\describe{
\item{a}{type of study program}
\item{b}{the particular school}
\item{act}{the individual's ACT score}
}
}

\section{Synonym}{

C10T5
}

\examples{
# Load the data
data(chapter_10_table_5)

# Or, alternatively load the data as
data(C10T5)

# View the structure
str(chapter_10_table_5)

}
\references{
Maxwell, S. E., Delaney, H. D., \& Kelley, K. (2018). \emph{Designing experiments and analyzing data:
{A} model comparison perspective} (3rd ed.). New York, NY: Routledge.
}
\author{
Ken Kelley \email{kkelley@nd.edu}
}
\keyword{datasets}
