% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{chapter_11_table_5}
\alias{chapter_11_table_5}
\alias{C11T5}
\alias{Chapter_11_Table_5}
\alias{c11t5}
\title{The data used in Chapter 11, Table 5}
\format{
An object of class \code{data.frame} with 12 rows and 4 columns.
}
\source{
\url{https://designingexperiments.com/data/}

Maxwell, S. E., Delaney, H. D., & Kelley, K. (2018). \emph{Designing experiments and
analyzing data: {A} model comparison perspective}. (3rd ed.). New York, NY: Routledge.
}
\usage{
data(chapter_11_table_5)
}
\description{
The data used in Chapter 11, Table 5
}
\details{
The data show that 12 participants have been observed in each of 4 conditions. To make the example easier to discuss, let's suppose that the 12 subjects are children who have been observed at 30, 36, 42, and 48 months of age. Essentially, for the present data set, 12 children were each observed four times over an 18 month period. The dependent variable is the age-normed general cognitive score on the McCarthy Scales of Children's Abilities. Interest is to determine if the children were sampled from a population where growth in cognitive ability is more rapid or less rapid than average.
}
\section{Variables}{

\describe{
\item{months30}{age-normed general cognitive score for 30-month-old}
\item{months36}{age-normed general cognitive score for 36-month-old}
\item{months42}{age-normed general cognitive score for 42-month-old}
\item{months48}{age-normed general cognitive score for 48-month-old}
}
}

\section{Synonym}{

C11T5
}

\examples{
# Load the data
data(chapter_11_table_5)

# Or, alternatively load the data as
data(C11T5)

# View the structure
str(chapter_11_table_5)

}
\references{
Maxwell, S. E., Delaney, H. D., \& Kelley, K. (2018). \emph{Designing experiments and analyzing data:
{A} model comparison perspective} (3rd ed.). New York, NY: Routledge.
}
\author{
Ken Kelley \email{kkelley@nd.edu}
}
\keyword{datasets}
