% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{chapter_14_table_10}
\alias{chapter_14_table_10}
\alias{C14T10}
\alias{Chapter_14_Table_10}
\alias{c14t10}
\title{The data used in Chapter 14, Table 10}
\format{
An object of class \code{data.frame} with 20 rows and 4 columns.
}
\source{
\url{https://designingexperiments.com/data/}

Maxwell, S. E., Delaney, H. D., & Kelley, K. (2018). \emph{Designing experiments and
analyzing data: {A} model comparison perspective}. (3rd ed.). New York, NY: Routledge.
}
\usage{
data(chapter_14_table_10)
}
\description{
The data used in Chapter 14, Table 10
}
\details{
These data are identical to those analyzed in chapter 12 (see Tables 12.7 and 12.15) to facilitate comparisons of the multivariate approach and the mixed-model approach.The hypothetical data contained in Table 14.10 contains an additional level of angle (four degrees) that was not considered in Table 14.8.
}
\section{Variables}{

\describe{
\item{angle0}{participant reaction time at angle 0}
\item{angle4}{participant reaction time at angle 4}
\item{angle8}{participant reaction time at angle 8}
\item{group}{participant group number}
}
}

\section{Synonym}{

C14T10
}

\examples{
# Load the data
data(chapter_14_table_10)

# Or, alternatively load the data as
data(C14T10)

# View the structure
str(chapter_14_table_10)

}
\references{
Maxwell, S. E., Delaney, H. D., \& Kelley, K. (2018). \emph{Designing experiments and analyzing data:
{A} model comparison perspective} (3rd ed.). New York, NY: Routledge.
}
\author{
Ken Kelley \email{kkelley@nd.edu}
}
\keyword{datasets}
