% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{chapter_4_table_1}
\alias{chapter_4_table_1}
\alias{C4T1}
\alias{Chapter_4_Table_1}
\alias{c4t1}
\title{The data used in Chapter 4, Table 1}
\format{
An object of class \code{data.frame} with 20 rows and 2 columns.
}
\source{
\url{https://designingexperiments.com/data/}

Maxwell, S. E., Delaney, H. D., & Kelley, K. (2018). \emph{Designing experiments and
analyzing data: {A} model comparison perspective}. (3rd ed.). New York, NY: Routledge.
}
\usage{
data(chapter_4_table_1)
}
\description{
The data used in Chapter 4, Table 1
}
\details{
This is hypothetical data for four groups of participants, corresponding to treatments for hypertension. The context is 24 mild hypertensives that have been independently and randomly assigned to one of the four treatments. The scores are the systolic blood pressure values two-weeks after the termination of treatment.
}
\section{Variables}{

\describe{
\item{\code{bloodpr}}{systolic blood pressure (hypothetical data)}
\item{\code{cond}}{identifies group membership (1=drug therapy; 2=biofeedbacbk; 3=diet; 4=combination)}}
}

\section{Synonym}{

C4T1
}

\examples{
# Load the data
data(chapter_4_table_1)

# Or, alternatively load the data as
data(C4T1)

# View the structure
str(chapter_4_table_1)

}
\references{
Maxwell, S. E., Delaney, H. D., \& Kelley, K. (2018). \emph{Designing experiments and analyzing data:
{A} model comparison perspective} (3rd ed.). New York, NY: Routledge.
}
\author{
Ken Kelley \email{kkelley@nd.edu}
}
\keyword{datasets}
