% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{chapter_5_table_4}
\alias{chapter_5_table_4}
\alias{C5T4}
\alias{Chapter_5_Table_4}
\alias{c5t4}
\title{The data used in Chapter 5, Table 4}
\format{
An object of class \code{data.frame} with 24 rows and 2 columns.
}
\source{
\url{https://designingexperiments.com/data/}

Maxwell, S. E., Delaney, H. D., & Kelley, K. (2018). \emph{Designing experiments and
analyzing data: {A} model comparison perspective}. (3rd ed.). New York, NY: Routledge.
}
\usage{
data(chapter_5_table_4)
}
\description{
The data used in Chapter 5, Table 4
}
\details{
The following data consists of blood pressure measurements for six individuals randomly assigned to one of four groups. Our purpose here is to perform four planed contrasts in order to discern if group differences exist for the selected contrasts of interests.
}
\section{Variables}{

\describe{
\item{group}{a numeric vector between 1 and 4;group number}
\item{sbp}{systolic blood pressure of a patient within one of the four groups}}
}

\section{Synonym}{

C5T4
}

\examples{
# Load the data
data(chapter_5_table_4)

# Or, alternatively load the data as
data(C5T4)

# View the structure
str(chapter_5_table_4)

}
\references{
Maxwell, S. E., Delaney, H. D., \& Kelley, K. (2018). \emph{Designing experiments and analyzing data:
{A} model comparison perspective} (3rd ed.). New York, NY: Routledge.
}
\author{
Ken Kelley \email{kkelley@nd.edu}
}
\keyword{datasets}
