% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{chapter_7_table_9}
\alias{chapter_7_table_9}
\alias{C7T9}
\alias{Chapter_7_Table_9}
\alias{c7t9}
\title{The data used in Chapter 7, Table 9}
\format{
An object of class \code{data.frame} with 36 rows and 3 columns.
}
\source{
\url{https://designingexperiments.com/data/}

Maxwell, S. E., Delaney, H. D., & Kelley, K. (2018). \emph{Designing experiments and
analyzing data: {A} model comparison perspective}. (3rd ed.). New York, NY: Routledge.
}
\usage{
data(chapter_7_table_9)
}
\description{
The data used in Chapter 7, Table 9
}
\details{
The following data is a generalization of the blood pressure data given in Table 7.5 (which itself was a generalization of the data given in Table 7.1). After the interaction is found to be significant, a common recommendation is to examine simple main effects. Recall that a simple main effect is the main effect of one factor given a fixed level of another factor. In this case interest is in determining if there are any differences in drugs (a) given biofeedback and (b) given no biofeedback.
}
\section{Variables}{

\describe{
\item{score}{blood pressure}
\item{feedback}{the likelihood of there being a biofeedback or drug main effect}
\item{drug}{the level of the drug factor}}
}

\section{Synonym}{

C7T9
}

\examples{
# Load the data
data(chapter_7_table_9)

# Or, alternatively load the data as
data(C7T9)

# View the structure
str(chapter_7_table_9)

}
\references{
Maxwell, S. E., Delaney, H. D., \& Kelley, K. (2018). \emph{Designing experiments and analyzing data:
{A} model comparison perspective} (3rd ed.). New York, NY: Routledge.
}
\author{
Ken Kelley \email{kkelley@nd.edu}
}
\keyword{datasets}
