% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{chapter_9_extension_exercise_2}
\alias{chapter_9_extension_exercise_2}
\alias{Chapter_9_Extension_Exercise_2}
\alias{C9ExtE2}
\alias{c9exte2}
\alias{chapter_9_extension_exercise_3}
\alias{Chapter_9_Extension_Exercise_3}
\alias{C9ExtE3}
\alias{c9exte3}
\title{The data used in Chapter 9 Extension, Exercise 2}
\format{
An object of class \code{data.frame} with 168 rows and 6 columns.

An object of class \code{data.frame} with 310 rows and 6 columns.
}
\source{
\url{https://designingexperiments.com/data/}

Maxwell, S. E., Delaney, H. D., & Kelley, K. (2018). \emph{Designing experiments and
analyzing data: {A} model comparison perspective}. (3rd ed.). New York, NY: Routledge.

\url{https://designingexperiments.com/data/}

Maxwell, S. E., Delaney, H. D., & Kelley, K. (2018). \emph{Designing experiments and
analyzing data: {A} model comparison perspective}. (3rd ed.). New York, NY: Routledge.
}
\usage{
data(chapter_9_extension_exercise_2)

data(chapter_9_extension_exercise_3)
}
\description{
Data from Chapter 9 Extension Exercise 2 of \emph{Designing Experiments and Analyzing Data: A Model Comparison Perspective} (3rd edition; Maxwell, Delaney, & Kelley).

Data from Chapter 9 Extension Exercise 3 of \emph{Designing Experiments and Analyzing Data: A Model Comparison Perspective} (3rd edition; Maxwell, Delaney, & Kelley).
}
\details{
\itemize{
\item ParticipantNumber.
\item Group.
\item BaseHrsDrkTypWk.
\item FolHrsDrkTypWk.
\item DiffBaseFolHrsDrk.
\item BaseHrsCtrd.}

\itemize{
\item Grade.
\item Treatment.
\item IQPre.
\item IQ4.
\item IQ8.
\item IQGain.}
}
\section{Synonym}{

C9ExtE2


C9ExtE3
}

\examples{
# Load the data
data(chapter_9_extension_exercise_2)

# Or, alternatively load the data as
data(C9ExtE2)

# View the structure
str(chapter_9_extension_exercise_2)

# Brief summary of the data.
summary(chapter_9_extension_exercise_2)

# Load the data
data(chapter_9_extension_exercise_3)

# Or, alternatively load the data as
data(C9ExtE3)

# View the structure
str(chapter_9_extension_exercise_3)

# Brief summary of the data.
summary(chapter_9_extension_exercise_2)

}
\references{
Maxwell, S. E., Delaney, H. D., & Kelley, K. (2018). \emph{Designing experiments and analyzing data:
{A} model comparison perspective} (3rd ed.). New York, NY: Routledge.

Maxwell, S. E., Delaney, H. D., & Kelley, K. (2018). \emph{Designing experiments and analyzing data:
{A} model comparison perspective} (3rd ed.). New York, NY: Routledge.
}
\author{
Ken Kelley \email{kkelley@nd.edu}

Ken Kelley \email{kkelley@nd.edu}
}
\keyword{datasets}
